% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{add_methods}
\alias{add_methods}
\title{Add model specification to fit}
\usage{
add_methods(x, engine)
}
\arguments{
\item{x}{A model fit}

\item{engine}{A character string for the model engine.}
}
\value{
The updated fit object.
}
\description{
Add model specification to fit
}
\keyword{internal}
