% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R, R/misc.R, R/print.R, R/update.R,
%   R/varying.R
\name{null_value}
\alias{null_value}
\alias{show_fit}
\alias{check_args}
\alias{update_dot_check}
\alias{new_model_spec}
\alias{check_outcome_levels}
\alias{check_final_param}
\alias{update_main_parameters}
\alias{update_engine_parameters}
\alias{print_model_spec}
\alias{update_spec}
\alias{is_varying}
\title{Functions required for parsnip-adjacent packages}
\usage{
null_value(x)

show_fit(model, eng)

check_args(object, call = rlang::caller_env())

update_dot_check(...)

new_model_spec(
  cls,
  args,
  eng_args,
  mode,
  user_specified_mode = TRUE,
  method,
  engine,
  user_specified_engine = TRUE,
  quantile_levels = NULL
)

check_outcome_levels(spec, y, call = rlang::caller_env())

check_final_param(x, call = rlang::caller_env())

update_main_parameters(args, param, call = rlang::caller_env())

update_engine_parameters(eng_args, fresh, ...)

print_model_spec(x, cls = class(x)[1], desc = get_model_desc(cls), ...)

update_spec(
  object,
  parameters,
  args_enquo_list,
  fresh,
  cls,
  ...,
  call = caller_env()
)

is_varying(x)
}
\description{
These functions are helpful when creating new packages that will register
new model specifications.
}
\keyword{internal}
