% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adds.R
\name{add_rowindex}
\alias{add_rowindex}
\title{Add a column of row numbers to a data frame}
\usage{
add_rowindex(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
The same data frame with a column of 1-based integers named \code{.row}.
}
\description{
Add a column of row numbers to a data frame
}
\examples{
\dontshow{if (!parsnip:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mtcars |> add_rowindex()
\dontshow{\}) # examplesIf}
}
