% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{as_matrix}
\alias{as_matrix}
\title{Convert data frame to matrix}
\usage{
as_matrix(x)
}
\arguments{
\item{x}{A data frame, matrix, or sparse matrix.}
}
\value{
A matrix or sparse matrix.
}
\description{
This is a substitute for \code{as.matrix()} that will convert a data frame to a
ordinary matrix but leave other formats (such as a sparse matrix) alone.
}
