% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_parsnip_spec.R
\name{attach_parsnip_models}
\alias{attach_parsnip_models}
\title{Attach model template objects to current environment}
\usage{
attach_parsnip_models(
  pattern = "*",
  mode = c("all", "regression", "classification"),
  env = rlang::global_env()
)
}
\arguments{
\item{pattern}{A character string for a regular expression used to filter
which specifications are attached.}

\item{mode}{One or more character strings for which collection of
specifications should be attached. Possible values are "all", "regression",
or "classification".}

\item{env}{An environment where the objects are attached.}
}
\value{
NULL (invisibly) but multiple objects are added to the chosen
environment as a side-effect.
}
\description{
\code{attach_parsnip_models()} will save a collection of \code{parsnip} model
specifications to the an environment (the global environment, by default).
These objects will have values of \verb{tune"()} in most commonly used augment
values (for use with the \code{tune} package). The naming convention is
\verb{\{md\}_\{function\}_\{engine\}} where \code{md} is an abbreviation of the model modes
(e.g., "reg", or "cls"), \code{function} is the name of the model specification
(e.g., "rand_forest"), and \code{engine} is the engine name.
}
