% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.model_fit}
\alias{augment.model_fit}
\title{Augment data with predictions}
\usage{
\method{augment}{model_fit}(x, new_data, ...)
}
\arguments{
\item{x}{A \code{model_fit} object produced by \code{\link[=fit]{fit()}} or \code{\link[=fit_xy]{fit_xy()}}.}

\item{new_data}{A data frame or matrix.}

\item{...}{Not currently used.}
}
\description{
\code{augment()} will add column(s) for predictions to the given data.
}
\details{
For regression models, a \code{.pred} column is added. If \code{x} was created using
\code{\link[=fit]{fit()}} and \code{new_data} contains the outcome column, a \code{.resid} column is
also added.

For classification models, the results include a column called \code{.pred_class}
as well as class probability columns named \verb{.pred_\{level\}}.
}
\examples{
reg_form <-
  linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ ., data = mtcars)
reg_xy <-
  linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit_xy(mtcars[, -1], mtcars$mpg)

augment(reg_form, head(mtcars))
augment(reg_form, head(mtcars[, -1]))

augment(reg_xy, head(mtcars))
augment(reg_xy, head(mtcars[, -1]))

# ------------------------------------------------------------------------------

data(two_class_dat, package = "modeldata")

cls_form <-
  logistic_reg() \%>\%
  set_engine("glm") \%>\%
  fit(Class ~ ., data = two_class_dat)
cls_xy <-
  logistic_reg() \%>\%
  set_engine("glm") \%>\%
  fit_xy(two_class_dat[, -3],
  two_class_dat$Class)

augment(cls_form, head(two_class_dat))
augment(cls_form, head(two_class_dat[, -3]))

augment(cls_xy, head(two_class_dat))
augment(cls_xy, head(two_class_dat[, -3]))

}
