% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart-internal}
\alias{bart-internal}
\alias{dbart_predict_calc}
\title{Developer functions for predictions via BART models}
\usage{
dbart_predict_calc(obj, new_data, type, level = 0.95, std_err = FALSE)
}
\arguments{
\item{obj}{A parsnip object.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{A single character value or \code{NULL}. Possible values are:
\itemize{
\item regression: "\code{numeric}"
\item classification: "\code{class}", "\code{prob}"
\item censored regression: "\code{survival}", "\code{time}", "\code{hazard}", "\code{linear_pred}"
\item quantile regression: "\code{quantile}"
\item interval estimates: "\code{conf_int}", "\code{pred_int}"
\item other:  "\code{raw}"
}

When \code{NULL}, \code{predict()} will choose an appropriate value based on the
model's mode.}

\item{level}{Confidence level.}

\item{std_err}{Attach column for standard error of prediction or not.}
}
\description{
Developer functions for predictions via BART models
}
\keyword{internal}
