% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_weights.R
\name{case_weights_allowed}
\alias{case_weights_allowed}
\title{Determine if case weights are used}
\usage{
case_weights_allowed(spec)
}
\arguments{
\item{spec}{A parsnip \link[=model_spec]{model specification}.}
}
\value{
A single logical.
}
\description{
Not all modeling engines can incorporate case weights into their
calculations. This function can determine whether they can be used.
}
\examples{
case_weights_allowed(linear_reg())
case_weights_allowed(linear_reg(engine = "keras"))
}
