% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_parsnip.R
\name{control_parsnip}
\alias{control_parsnip}
\alias{fit_control}
\title{Control the fit function}
\usage{
control_parsnip(verbosity = 1L, catch = FALSE)

fit_control(verbosity = 1L, catch = FALSE)
}
\arguments{
\item{verbosity}{An integer where a value of zero indicates
that no messages or output should be shown when packages are
loaded or when the model is fit. A value of 1 means that package
loading is quiet but model fits can produce output to the screen
(depending on if they contain their own \code{verbose}-type
argument). A value of 2 or more indicates that any output should
be seen.}

\item{catch}{A logical where a value of \code{TRUE} will evaluate
the model inside of \code{try(, silent = TRUE)}. If the model fails,
an object is still returned (without an error) that inherits the
class "try-error".}
}
\value{
An S3 object with class "fit_control" that is a named list with the
results of the function call
}
\description{
Options can be passed to the \code{\link[=fit]{fit()}} function that control the output and
computations
}
\details{
\code{fit_control()} is deprecated in favor of \code{control_parsnip()}.
}
