% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{convert_args}
\alias{convert_args}
\alias{get_arg_defaults}
\alias{get_arg}
\title{Find args for documentation}
\usage{
convert_args(model_name)

get_arg_defaults(model)

get_arg(ns, f, arg)
}
\description{
Find args for documentation
}
\keyword{internal}
