% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_weights.R
\name{convert_case_weights}
\alias{convert_case_weights}
\alias{convert_case_weights.hardhat_importance_weights}
\alias{convert_case_weights.hardhat_frequency_weights}
\title{Convert case weights to final from}
\usage{
convert_case_weights(x, where = "parsnip", ...)

\method{convert_case_weights}{hardhat_importance_weights}(x, where = "parsnip", ...)

\method{convert_case_weights}{hardhat_frequency_weights}(x, where = "parsnip", ...)
}
\arguments{
\item{x}{A vector with class \code{"hardhat_case_weights"}.}

\item{where}{The location where they will be used: \code{"parsnip"} or
\code{"yardstick"}.}
}
\value{
A numeric vector or NULL.
}
\description{
tidymodels requires case weights to have special classes. To use them in
model fitting or performance evaluation, they need to be converted to
numeric.
}
