% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{convert_form_to_xy_fit}
\alias{convert_form_to_xy_fit}
\alias{convert_form_to_xy_new}
\title{Helper functions to convert from formula to matrix interface}
\usage{
convert_form_to_xy_fit(
  formula,
  data,
  ...,
  na.action = na.omit,
  indicators = "traditional",
  composition = "data.frame",
  remove_intercept = TRUE
)

convert_form_to_xy_new(
  object,
  new_data,
  na.action = na.pass,
  composition = "data.frame"
)
}
\arguments{
\item{formula}{An object of class \code{formula} (or one that can
be coerced to that class): a symbolic description of the model
to be fitted.}

\item{data}{A data frame containing all relevant variables (e.g. outcome(s),
predictors, case weights, etc).}

\item{...}{Additional arguments passed to \code{\link[stats:model.frame]{stats::model.frame()}} and
specification of \code{offset} and \code{contrasts}.}

\item{na.action}{A function which indicates what should happen when the data
contain NAs.}

\item{indicators}{A string describing whether and how to create
indicator/dummy variables from factor predictors.}

\item{composition}{A string describing whether the resulting \code{x} and \code{y}
should be returned as a \code{"matrix"} or a \code{"data.frame"}.}

\item{remove_intercept}{A logical indicating whether to remove the intercept
column after \code{model.matrix()} is finished.}

\item{object}{An object of class \code{model_fit}.}
}
\description{
Functions to take a formula interface and get the resulting
objects (y, x, weights, etc) back. For the most part, this
emulates the internals of \code{lm} (and also see the notes at
https://developer.r-project.org/model-fitting-functions.html).
}
\details{
\code{convert_form_to_xy_fit} is for when the data are created for modeling.
It saves both the data objects as well as the objects needed when new data
are predicted (e.g. \code{terms}, etc.).

\code{convert_form_to_xy_new} is used when new samples are being predicted and
only requires the predictors to be available.
}
\keyword{internal}
