% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{convert_stan_interval}
\alias{convert_stan_interval}
\title{Convenience function for intervals}
\usage{
convert_stan_interval(x, level = 0.95, lower = TRUE)
}
\arguments{
\item{x}{A fitted model object}

\item{level}{Level of uncertainty for intervals}

\item{lower}{Is \code{level} the lower level?}
}
\description{
Convenience function for intervals
}
\keyword{internal}
