% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-reg-doc-lm.R
\name{details-linear-reg-lm}
\alias{details-linear-reg-lm}
\title{Linear regression via lm}
\description{
\code{\link[stats:lm]{stats::lm()}} uses ordinary least squares to fit models with numeric outcomes.
}
\details{
For this engine, there is a single mode: regression
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package}{\if{html}{\out{<div class="r">}}\preformatted{linear_reg() \%>\% 
  set_engine("lm") \%>\% 
  set_mode("regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Linear Regression Model Specification (regression)
## 
## Computational engine: lm 
## 
## Model fit template:
## stats::lm(formula = missing_arg(), data = missing_arg(), weights = missing_arg())
}
}

\subsection{Preprocessing requirements}{

The only preprocessing required for this model is that
factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables). When using the formula method via
\code{\link[=fit.model_spec]{fit.model_spec()}}, \code{parsnip} will convert these
data if required.
}

\subsection{Other details}{

The model can accept case weights.
}

\subsection{References}{
\itemize{
\item Kuhn, M, and K Johnson. 2013. \emph{Applied Predictive Modeling}.
Springer.
}
}
}
\keyword{internal}
