% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_additive_mod_vgam.R
\name{details_gen_additive_mod_vgam}
\alias{details_gen_additive_mod_vgam}
\title{Ordinal generalized additive model via vector GAMs}
\description{
\code{\link[VGAM:vgam]{VGAM::vgam()}} fits vector generalized additive models, which specialize to
several families of ordinal regression models.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

There are no main tuning parameters for this model. Two relevant engine
parameters are:
\itemize{
\item \code{link}: the link function such as \code{logistic}, \code{probit}, \code{loglog},
\code{cloglog}, or \code{cauchit}.
\item \code{family}: the function to contrast levels such as \code{cumulative_link},
\code{adjacent_categories}, \code{continuation_ratio}, or \code{stopping_ratio}
}
}

\subsection{Translation from parsnip to the original package}{

The \strong{ordered} extension package is required to fit this model.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(ordered)

gen_additive_mod() \%>\% 
  set_engine("vgam") \%>\% 
  set_mode("classification") \%>\% 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## GAM Model Specification (classification)
## 
## Computational engine: vgam 
## 
## Model fit template:
## ordered::VGAM_vgam_wrapper(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg(), parallel = TRUE)
}\if{html}{\out{</div>}}
}

\subsection{Model fitting}{

This model should be used with a model formula so that smooth terms can
be specified. For example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(VGAM)
# Make number of cylinders and ordered factor
ord_cars <- mtcars[, -1]
ord_cars$cyl <- as.ordered(ord_cars$cyl)

car_fit <- 
  gen_additive_mod() |> 
  set_engine("vgam") |> 
  set_mode("classification") |> 
  fit(cyl ~ disp + s(wt) + am, data = ord_cars)
}\if{html}{\out{</div>}}

The smoothness of the terms will need to be manually specified (e.g.,
using \code{s(x, df = 10)}) in the formula.

When using a workflow, pass the \emph{model formula} to
\code{\link[workflows:add_model]{workflows::add_model()}}’s \code{formula} argument,
and a simplified \emph{preprocessing formula} elsewhere.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{spec <- 
  gen_additive_mod() |> 
  set_engine("vgam") |> 
  set_mode("classification")

workflow() |> 
  add_model(spec, formula = cyl ~ disp + s(wt) + am) |> 
  add_formula(cyl ~ disp + wt + am) |> 
  fit(data = ord_cars) |> 
  extract_fit_engine()
}\if{html}{\out{</div>}}

To learn more about the differences between these formulas, see
\code{\link[=model_formula]{?model_formula}}.
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine. When using the
formula method via \code{\link[=fit.model_spec]{fit()}}, parsnip will
convert factor columns to indicators.
}

\subsection{Case weights}{

This model can utilize case weights during model fitting. To use them,
see the documentation in \link{case_weights} and the examples
on \code{tidymodels.org}.

The \code{fit()} and \code{fit_xy()} functions have arguments called
\code{case_weights} that expect vectors of case weights.
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("gen_additive_mod_predict") |>
  dplyr::filter(engine == "vgam") |> 
  dplyr::select(mode, type)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 2 x 2
##   mode           type 
##   <chr>          <chr>
## 1 classification class
## 2 classification prob
}\if{html}{\out{</div>}}
}

\subsection{References}{
\itemize{
\item Yee T. 2010. The VGAM Package for Categorical Data Analysis. \emph{Journal
of Statistical Software} 32(10):1–34. .
}
}
}
\keyword{internal}
