% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_brulee.R
\name{details_mlp_brulee}
\alias{details_mlp_brulee}
\title{Multilayer perceptron via brulee}
\description{
\code{\link[brulee:brulee_mlp]{brulee::brulee_mlp()}} fits a neural networks.
}
\details{
For this engine, there are multiple modes: classification and regression
\subsection{Tuning Parameters}{\preformatted{## Error: Tibble columns must have compatible sizes.
## * Size 6: Existing data.
## * Size 8: Column `default`.
## ℹ Only values of size one are recycled.

## Error in is.data.frame(y): object 'defaults' not found
}
}
}
\keyword{internal}
