% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_qrnn.R
\name{details_mlp_qrnn}
\alias{details_mlp_qrnn}
\title{Multilayer perceptron for quantile regression via qrnn}
\description{
\code{\link[qrnn:mcqrnn.fit]{qrnn::mcqrnn.fit()}} fits a neural network for quantile regression.
}
\details{
For this engine, there is a single mode: quantile regression
\subsection{Tuning Parameters}{

This model has 5 tuning parameters:
\itemize{
\item \code{hidden_units}: # Hidden Units (type: integer, default: 2L)
\item \code{penalty}: Amount of Regularization (type: double, default: 0.0)
\item \code{epochs}: # Epochs (type: integer, default: 5000L)
\item \code{activation}: Activation Function (type: character, default:
‘sigmoid’)
\item \code{learn_rate}: Learning Rate (type: double, default: 0.01)
}

\emph{Note}: the learning rate is only used when \code{method = "adam"}.

Other engine arguments of interest:
\itemize{
\item \code{n.trials}: number of repeated trials used to avoid local minima.
\item \code{method}: The optimization technique (\code{"nlm"} or \code{"adam"}).
\item \code{iterbreak}: the maximum number of ADAM iterations without progress
before the optimization is stopped.
\item \code{minibatch}: the batch size for ADAM.
}
}

\subsection{Translation from parsnip to the original package (quantile regression)}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{mlp(
  hidden_units = integer(1),
  penalty = double(1),
  epochs = integer(1),
  activation = character(1),
  learn_rate = double(1)
) |>  
  set_engine("qrnn") |> 
  set_mode("quantile regression", quantile_levels = (1:3) / 4) |> 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Single Layer Neural Network Model Specification (quantile regression)
## 
## Main Arguments:
##   hidden_units = integer(1)
##   penalty = double(1)
##   epochs = integer(1)
##   activation = character(1)
##   learn_rate = double(1)
## 
## Computational engine: qrnn 
## 
## Model fit template:
## parsnip::mcqrnn_train(x = missing_arg(), y = missing_arg(), n.hidden = integer(1), 
##     penalty = double(1), iter.max = integer(1), Th = character(1), 
##     alpha = double(1), trace = FALSE, tau = quantile_levels)

## Quantile levels: 0.25, 0.5, and 0.75.
}\if{html}{\out{</div>}}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine. When using the
formula method via \code{\link[=fit.model_spec]{fit()}}, parsnip will
convert factor columns to indicators.

Predictors should have the same scale. One way to achieve this is to
center and scale each so that each predictor has mean zero and a
variance of one.
}

\subsection{Case weights}{

The underlying model implementation does not allow for case weights.
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("mlp_predict") |>
  dplyr::filter(engine == "qrnn") |>
  dplyr::select(mode, type)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 1 x 2
##   mode                type    
##   <chr>               <chr>   
## 1 quantile regression quantile
}\if{html}{\out{</div>}}
}

\subsection{References}{
\itemize{
\item Cannon, A.J., 2018. Non-crossing nonlinear regression quantiles by
monotone composite quantile regression neural network, with
application to rainfall extremes. \emph{Stochastic Environmental Research
and Risk Assessment}, 32(11): 3207-3225.
\url{doi:10.1007/s00477-018-1573-6}
}
}
}
\keyword{internal}
