% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_reg_h2o.R
\name{details_multinom_reg_h2o}
\alias{details_multinom_reg_h2o}
\title{Multinomial regression via h2o}
\description{
\code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} fits a model that uses linear predictors to predict
multiclass data for multinomial responses.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 2 tuning parameters:
\itemize{
\item \code{mixture}: Proportion of Lasso Penalty (type: double, default: see
below)
\item \code{penalty}: Amount of Regularization (type: double, default: see below)
}

By default, when not given a fixed \code{penalty},
\code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} uses a heuristic approach to select
the optimal value of \code{penalty} based on training data. Setting the
engine parameter \code{lambda_search} to \code{TRUE} enables an efficient version
of the grid search, see more details at
\url{https://docs.h2o.ai/h2o/latest-stable/h2o-docs/data-science/algo-params/lambda_search.html}.

The choice of \code{mixture} depends on the engine parameter \code{solver}, which
is automatically chosen given training data and the specification of
other model parameters. When \code{solver} is set to \code{'L-BFGS'}, \code{mixture}
defaults to 0 (ridge regression) and 0.5 otherwise.
}

\subsection{Translation from parsnip to the original package}{

\code{\link[agua:h2o_train_glm]{agua::h2o_train_glm()}} for \code{multinom_reg()} is
a wrapper around \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} with
\code{family = 'multinomial'}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{multinom_reg(penalty = double(1), mixture = double(1)) |> 
  set_engine("h2o") |> 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Multinomial Regression Model Specification (classification)
## 
## Main Arguments:
##   penalty = double(1)
##   mixture = double(1)
## 
## Computational engine: h2o 
## 
## Model fit template:
## agua::h2o_train_glm(x = missing_arg(), y = missing_arg(), weights = missing_arg(), 
##     validation_frame = missing_arg(), lambda = double(1), alpha = double(1), 
##     family = "multinomial")
}\if{html}{\out{</div>}}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine. When using the
formula method via \code{\link[=fit.model_spec]{fit()}}, parsnip will
convert factor columns to indicators.

Predictors should have the same scale. One way to achieve this is to
center and scale each so that each predictor has mean zero and a
variance of one.

By default, \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} uses the argument
\code{standardize = TRUE} to center and scale the data.
}

\subsection{Case weights}{

This model can utilize case weights during model fitting. To use them,
see the documentation in \link{case_weights} and the examples
on \code{tidymodels.org}.

The \code{fit()} and \code{fit_xy()} functions have arguments called
\code{case_weights} that expect vectors of case weights.
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("multinom_reg_predict") |>
  dplyr::filter(engine == "h2o") |>
  dplyr::select(mode, type)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 2 x 2
##   mode           type 
##   <chr>          <chr>
## 1 classification class
## 2 classification prob
}\if{html}{\out{</div>}}
}

\subsection{Initializing h2o}{

To use the h2o engine with tidymodels, please run \code{h2o::h2o.init()}
first. By default, This connects R to the local h2o server. This needs
to be done in every new R session. You can also connect to a remote h2o
server with an IP address, for more details see
\code{\link[h2o:h2o.init]{h2o::h2o.init()}}.

You can control the number of threads in the thread pool used by h2o
with the \code{nthreads} argument. By default, it uses all CPUs on the host.
This is different from the usual parallel processing mechanism in
tidymodels for tuning, while tidymodels parallelizes over resamples, h2o
parallelizes over hyperparameter combinations for a given resample.

h2o will automatically shut down the local h2o instance started by R
when R is terminated. To manually stop the h2o server, run
\code{h2o::h2o.shutdown()}.
}
}
\keyword{internal}
