% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_Bayes_h2o.R
\name{details_naive_Bayes_h2o}
\alias{details_naive_Bayes_h2o}
\title{Naive Bayes models via naivebayes}
\description{
\code{\link[h2o:h2o.naiveBayes]{h2o::h2o.naiveBayes()}} fits a model that uses Bayes' theorem to compute
the probability of each class, given the predictor values.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 1 tuning parameter:
\itemize{
\item \code{Laplace}: Laplace Correction (type: double, default: 0.0)
}

\code{\link[h2o:h2o.naiveBayes]{h2o::h2o.naiveBayes()}} provides several engine
arguments to deal with imbalances and rare classes:
\itemize{
\item \code{balance_classes} A logical value controlling over/under-sampling (for
imbalanced data). Defaults to \code{FALSE}.
\item \code{class_sampling_factors} The over/under-sampling ratios per class (in
lexicographic order). If not specified, sampling factors will be
automatically computed to obtain class balance during training.
Require \code{balance_classes} to be \code{TRUE}.
\item \code{min_sdev}: The minimum standard deviation to use for observations
without enough data, must be greater than 1e-10.
\item \code{min_prob}: The minimum probability to use for observations with not
enough data.
}
}

\subsection{Translation from parsnip to the original package}{

The \strong{agua} extension package is required to fit this model.

\code{\link[agua:h2o_train_nb]{agua::h2o_train_nb()}} is a wrapper around
\code{\link[h2o:h2o.naiveBayes]{h2o::h2o.naiveBayes()}}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(agua)

naive_Bayes(Laplace = numeric(0)) |>
  set_engine("h2o") |> 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Naive Bayes Model Specification (classification)
## 
## Main Arguments:
##   Laplace = numeric(0)
## 
## Computational engine: h2o 
## 
## Model fit template:
## agua::h2o_train_nb(x = missing_arg(), y = missing_arg(), weights = missing_arg(), 
##     validation_frame = missing_arg(), laplace = numeric(0))
}\if{html}{\out{</div>}}
}

\subsection{Case weights}{

This model can utilize case weights during model fitting. To use them,
see the documentation in \link{case_weights} and the examples
on \code{tidymodels.org}.

The \code{fit()} and \code{fit_xy()} functions have arguments called
\code{case_weights} that expect vectors of case weights.
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("naive_Bayes_predict") |>
  dplyr::filter(engine == "h2o") |>
  dplyr::select(mode, type)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 2 x 2
##   mode           type 
##   <chr>          <chr>
## 1 classification class
## 2 classification prob
}\if{html}{\out{</div>}}
}

\subsection{Initializing h2o}{

To use the h2o engine with tidymodels, please run \code{h2o::h2o.init()}
first. By default, This connects R to the local h2o server. This needs
to be done in every new R session. You can also connect to a remote h2o
server with an IP address, for more details see
\code{\link[h2o:h2o.init]{h2o::h2o.init()}}.

You can control the number of threads in the thread pool used by h2o
with the \code{nthreads} argument. By default, it uses all CPUs on the host.
This is different from the usual parallel processing mechanism in
tidymodels for tuning, while tidymodels parallelizes over resamples, h2o
parallelizes over hyperparameter combinations for a given resample.

h2o will automatically shut down the local h2o instance started by R
when R is terminated. To manually stop the h2o server, run
\code{h2o::h2o.shutdown()}.
}

\subsection{Saving fitted model objects}{

Models fitted with this engine may require native serialization methods
to be properly saved and/or passed between R sessions. To learn more
about preparing fitted models for serialization, see the bundle package.
}
}
\keyword{internal}
