% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_reg_polr.R
\name{details_ordinal_reg_polr}
\alias{details_ordinal_reg_polr}
\title{Ordinal regression via polr}
\description{
\code{\link[MASS:polr]{MASS::polr()}} fits a cumulative-link ordinal regression model.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package}{

The \strong{ordered} extension package is required to fit this model.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(ordered)

ordinal_reg() \%>\% 
  set_engine("polr") \%>\% 
  set_mode("classification") \%>\% 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Ordinal Regression Model Specification (classification)
## 
## Computational engine: polr 
## 
## Model fit template:
## MASS::polr(formula = missing_arg(), data = missing_arg(), weights = missing_arg())
}\if{html}{\out{</div>}}
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("ordinal_reg_predict") |>
  dplyr::filter(engine == "polr") |>
  dplyr::select(mode, type)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 2 x 2
##   mode           type 
##   <chr>          <chr>
## 1 classification class
## 2 classification prob
}\if{html}{\out{</div>}}
}

\subsection{References}{
\itemize{
\item Bürkner P-C, Vuorre M. 2019. Ordinal Regression Models in Psychology:
A Tutorial. \emph{Advances in Methods and Practices in Psychological
Science} 2(1):77-101. .
}
}
}
\keyword{internal}
