% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_reg_vgam.R
\name{details_ordinal_reg_vgam}
\alias{details_ordinal_reg_vgam}
\title{Ordinal regression via vector GAMs}
\description{
\code{\link[VGAM:vgam]{VGAM::vgam()}} fits vector generalized additive models, which specialize to
several families of ordinal regression models.
}
\keyword{internal}
