% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_reg_flexsurv.R
\name{details_surv_reg_flexsurv}
\alias{details_surv_reg_flexsurv}
\title{Parametric survival regression}
\description{
\code{\link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg()}} fits a parametric survival model.
}
\details{
For this engine, there is a single mode: regression
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{surv_reg() \%>\% 
  set_engine("flexsurv") \%>\% 
  set_mode("regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Warning: `surv_reg()` was deprecated in parsnip 0.1.6.
## Please use `survival_reg()` instead.

## Parametric Survival Regression Model Specification (regression)
## 
## Computational engine: flexsurv 
## 
## Model fit template:
## flexsurv::flexsurvreg(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg())
}
}

\subsection{References}{
\itemize{
\item Jackson, C. 2016. \code{flexsurv}: A Platform for Parametric Survival
Modeling in R. \emph{Journal of Statistical Software}, 70(8), 1 - 33.
}
}
}
\keyword{internal}
