% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_reg_survival.R
\name{details_surv_reg_survival}
\alias{details_surv_reg_survival}
\title{Parametric survival regression}
\description{
\code{\link[survival:survreg]{survival::survreg()}} fits a parametric survival model.
}
\details{
For this engine, there is a single mode: regression
\subsection{Tuning Parameters}{

This model has no tuning parameters.
}

\subsection{Translation from parsnip to the original package}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{surv_reg() \%>\% 
  set_engine("survival") \%>\% 
  set_mode("regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Warning: `surv_reg()` was deprecated in parsnip 0.1.6.
## Please use `survival_reg()` instead.

## Parametric Survival Regression Model Specification (regression)
## 
## Computational engine: survival 
## 
## Model fit template:
## survival::survreg(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg(), model = TRUE)
}
}

\subsection{References}{
\itemize{
\item Kalbfleisch, J. D. and Prentice, R. L. 2002 \emph{The statistical
analysis of failure time data}, Wiley.
}
}
}
\keyword{internal}
