% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_rln_brulee.R
\name{details_tabular_rln_brulee}
\alias{details_tabular_rln_brulee}
\title{Regularization learning network via brulee}
\description{
\code{\link[brulee:brulee_rln]{brulee::brulee_rln()}} fits a regularization learning network.
}
\details{
For this engine, there are multiple modes:
\subsection{Tuning Parameters}{

This model has 0 tuning parameters:

\code{penalty_average} and \code{step_rate} are specified on the natural scale but
are best tuned on the log10 scale.
}

\subsection{Translation from parsnip to the original package (regression)}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tabular_rln(
  hidden_units = integer(1),
  penalty_type = character(1),
  penalty_average = double(1),
  step_rate = double(1),
  activation = character(1),
  epochs = integer(1),
  learn_rate = double(1),
  rate_schedule = character(1),
  momentum = double(1),
  batch_size = NULL,
  stop_iter = integer(1)
) |>
  set_engine("brulee") |>
  set_mode("regression") |>
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## ! parsnip could not locate an implementation for `tabular_rln` regression model
##   specifications using the `brulee` engine.
## i The parsnip extension package tabby implements support for this specification.
## i Please install (if needed) and load to continue.

## tabular rln Model Specification (regression)
## 
## Main Arguments:
##   hidden_units = integer(1)
##   penalty_type = character(1)
##   penalty_average = double(1)
##   step_rate = double(1)
##   activation = character(1)
##   epochs = integer(1)
##   learn_rate = double(1)
##   rate_schedule = character(1)
##   momentum = double(1)
##   stop_iter = integer(1)
## 
## Computational engine: brulee
}\if{html}{\out{</div>}}
}

\subsection{Preprocessing requirements}{

\code{brulee_rln()} requires numeric predictors. Factor or categorical
predictors need to be converted to numeric values (e.g., dummy or
indicator variables) before fitting; parsnip does \emph{not} create indicator
variables for this engine, so use a recipe (or some other method) to
make them numeric.

Predictors should have the same scale. One way to achieve this is to
center and scale each so that each predictor has mean zero and a
variance of one.
}

\subsection{Case weights}{

The underlying model implementation does not allow for case weights.
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("tabular_rln_predict") |>
  dplyr::filter(engine == "brulee") |>
  dplyr::select(mode, type)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 0 x 2
## # i 2 variables: mode <chr>, type <chr>
}\if{html}{\out{</div>}}
}

\subsection{References}{
\itemize{
\item Shavitt, I., & Segal, E. (2018). Regularization learning networks:
Deep learning for tabular datasets. \emph{Advances in Neural Information
Processing Systems}, 31, 1379-1389.
}
}
}
\keyword{internal}
