% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_models.R
\name{doc-tools}
\alias{doc-tools}
\alias{find_engine_files}
\title{Tools for documenting packages}
\usage{
find_engine_files(mod)
}
\arguments{
\item{mod}{A character string for the model file}

\item{pkg}{The package that contains the model file}
}
\value{
\code{find_engine_files()} returns a character string.
}
\description{
Tools for documenting packages
}
\examples{
cat(find_engine_files("linear_reg"))
}
\keyword{internal}
