% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helpers.R
\name{.extract_surv_status}
\alias{.extract_surv_status}
\title{Extract survival status}
\usage{
.extract_surv_status(surv)
}
\arguments{
\item{surv}{A single \code{\link[survival:Surv]{survival::Surv()}} object.}
}
\value{
A numeric vector.
}
\description{
Extract the status from a \code{\link[survival:Surv]{survival::Surv()}} object.
}
