% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helpers.R
\name{.time_as_binary_event}
\alias{.time_as_binary_event}
\title{Convert survival objects to binary factors}
\arguments{
\item{surv}{A single \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{eval_time}{A single numeric value for the evaluation time.}
}
\value{
A two level factor.
}
\description{
For a given evaluation time, convert a \code{\link[survival:Surv]{survival::Surv()}} object to a binary
factor with levels \code{"event"} and \code{"non-event"}.
}
\details{
The following three cases can occur:
\itemize{
\item \strong{Events}: Evaluation time is greater than or equal to the event time
("it has already happened").
\item \strong{Non-events}: Evaluation time is less than the observed time, censored
or not ("nothing has happened yet").
\item \strong{Ambiguous outcomes}: Evaluation time is greater than or equal to the
observed censored time ("we don't know if anything might have happened by now").
A missing value is returned for these observations.
}
}
\references{
Graf, E., Schmoor, C., Sauerbrei, W., and Schumacher, M. (1999).
"Assessment and Comparison of Prognostic Classification Schemes for Survival Data."
\emph{Statistics in Medicine}, 18, 2529-2545.
}
