% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{eval_args}
\alias{eval_args}
\title{Evaluate parsnip model arguments}
\usage{
eval_args(spec, ...)
}
\arguments{
\item{spec}{A \link[=model_spec]{model specification}.}

\item{...}{Not used.}
}
\description{
Evaluate parsnip model arguments
}
\keyword{internal}
