% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalize.R
\name{finalize}
\alias{finalize}
\title{Resolve a Model Specification for a Computational Engine}
\usage{
finalize(x, ...)
}
\arguments{
\item{x}{A model specification.}

\item{...}{Not currently used.}
}
\description{
\code{finalize} will translate a model specification into a code
object that is specific to a particular engine (e.g. R package).
It translates generic parameters to their counterparts.
}
