% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_models.R
\name{get_model_env}
\alias{get_model_env}
\alias{set_new_model}
\alias{set_model_mode}
\alias{set_model_engine}
\alias{set_model_arg}
\alias{set_dependency}
\alias{get_dependency}
\alias{set_fit}
\alias{get_fit}
\alias{set_pred}
\alias{get_pred_type}
\alias{show_model_info}
\title{Tools to Register Models}
\usage{
get_model_env()

set_new_model(mod)

set_model_mode(mod, mode)

set_model_engine(mod, mode, eng)

set_model_arg(mod, eng, val, original, func, submodels)

set_dependency(mod, eng, pkg)

get_dependency(mod)

set_fit(mod, mode, eng, value)

get_fit(mod)

set_pred(mod, mode, eng, type, value)

get_pred_type(mod, type)

show_model_info(mod)
}
\description{
Tools to Register Models
}
\keyword{internal}
