% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-engines.R
\name{.check_glmnet_penalty_fit}
\alias{.check_glmnet_penalty_fit}
\alias{.check_glmnet_penalty_predict}
\title{Helper functions for checking the penalty of glmnet models}
\usage{
.check_glmnet_penalty_fit(x, engine, call = rlang::caller_env())

.check_glmnet_penalty_predict(
  penalty = NULL,
  object,
  multi = FALSE,
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{An object of class \code{model_spec}.}

\item{penalty}{A penalty value to check.}

\item{object}{An object of class \code{model_fit}.}

\item{multi}{A logical indicating if multiple values are allowed.}
}
\description{
These functions are for developer use.

\code{.check_glmnet_penalty_fit()} checks that the model specification for fitting a
glmnet model contains a single value.

\code{.check_glmnet_penalty_predict()} checks that the penalty value used for prediction is valid.
If called by \code{predict()}, it needs to be a single value. Multiple values are
allowed for \code{multi_predict()}.
}
\keyword{internal}
