% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet-engines.R
\name{.organize_glmnet_pred}
\alias{.organize_glmnet_pred}
\title{Organize glmnet predictions}
\usage{
.organize_glmnet_pred(x, object)
}
\arguments{
\item{x}{Predictions as returned by the \code{predict()} method for glmnet models.}

\item{object}{An object of class \code{model_fit}.}
}
\description{
This function is for developer use and organizes predictions from glmnet
models.
}
\keyword{internal}
