% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_h2o_helpers.R
\name{h2o_train}
\alias{h2o_train}
\title{Modeling via h2o}
\usage{
h2o_train(x, y, h2o_fun = c("h2o.randomForest"), ...)
}
\arguments{
\item{x}{A data frame or matrix of predictors}

\item{y}{A vector (factor or numeric) or matrix (numeric) of outcome data.}

\item{h2o_fun}{The name of the h2o modeling function.}

\item{...}{Other options to pass to \code{fun}.}
}
\value{
A fitted \code{h2o} object.
}
\description{
\code{h2o_train} is a wrapper for \code{h2o} models to start an H2O cluster,
convert data.frame input to H2OFrames and
where all of the model arguments are in the main function.
}
