% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{has_multi_pred}
\alias{has_multi_pred}
\alias{has_multi_pred.default}
\alias{has_multi_pred.model_fit}
\alias{has_multi_pred.workflow}
\title{Determine if a model can make predictions on sub-models}
\usage{
has_multi_pred(object, ...)

\method{has_multi_pred}{default}(object, ...)

\method{has_multi_pred}{model_fit}(object, ...)

\method{has_multi_pred}{workflow}(object, ...)
}
\arguments{
\item{object}{An object to test.}

\item{...}{Not currently used.}
}
\value{
A single logical value.
}
\description{
Determine if a model can make predictions on sub-models
}
\examples{
model_idea <- linear_reg() \%>\% set_engine("lm")
has_multi_pred(model_idea)
model_fit <- fit(model_idea, mpg ~ ., data = mtcars)
has_multi_pred(model_fit)
}
\keyword{internal}
