% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras3.R
\name{keras3_activations}
\alias{keras3_activations}
\title{Activation functions for neural networks in keras3}
\usage{
keras3_activations()
}
\value{
A character vector of values.
}
\description{
Activation functions for neural networks in keras3
}
\keyword{internal}
