% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras3.R
\name{keras3_mlp}
\alias{keras3_mlp}
\title{Simple interface to MLP models via keras3}
\usage{
keras3_mlp(
  x,
  y,
  hidden_units = 5,
  penalty = 0,
  dropout = 0,
  epochs = 20,
  activation = "softmax",
  seed = sample.int(10^5, size = 1),
  ...
)
}
\arguments{
\item{x}{A data frame or matrix of predictors.}

\item{y}{A vector (factor or numeric) or matrix (numeric) of outcome data.}

\item{hidden_units}{An integer for the number of hidden units.}

\item{penalty}{A non-negative real number for the amount of weight decay.
Either this parameter \emph{or} \code{dropout} can be specified.}

\item{dropout}{The proportion of parameters to set to zero. Either this
parameter \emph{or} \code{penalty} can be specified.}

\item{epochs}{An integer for the number of passes through the data.}

\item{activation}{A character string for the type of activation function
between layers.}

\item{seed}{A single positive integer to control randomness.}

\item{...}{Additional named arguments to pass to \code{keras3::compile()} or
\code{keras3::fit()}.}
}
\value{
A \code{keras3} model object.
}
\description{
Instead of building a \code{keras3} model sequentially, \code{keras3_mlp} can be used
to create a feedforward network with a single hidden layer. Regularization
is via either weight decay or dropout.
}
\keyword{internal}
