% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras3.R
\name{keras3_predict_classes}
\alias{keras3_predict_classes}
\title{Wrapper for keras3 class predictions}
\usage{
keras3_predict_classes(object, x)
}
\arguments{
\item{object}{A keras3 model fit}

\item{x}{A data set.}
}
\description{
Wrapper for keras3 class predictions
}
\keyword{internal}
