% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{keras_activations}
\alias{keras_activations}
\title{Activation functions for neural networks in keras}
\usage{
keras_activations()
}
\value{
A character vector of values.
}
\description{
Activation functions for neural networks in keras
}
\keyword{internal}
