% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp_data.R
\name{keras_mlp}
\alias{keras_mlp}
\title{MLP in Keras}
\usage{
keras_mlp(x, y, hidden_units = 5, decay = 0, dropout = 0,
  epochs = 20, act = "softmax", seeds = sample.int(10^5, size = 3),
  ...)
}
\description{
MLP in Keras
}
