% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{keras_predict_classes}
\alias{keras_predict_classes}
\title{Wrapper for keras class predictions}
\usage{
keras_predict_classes(object, x)
}
\arguments{
\item{object}{A keras model fit}

\item{x}{A data set.}
}
\description{
Wrapper for keras class predictions
}
\keyword{internal}
