% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{keras_set_seed}
\alias{keras_set_seed}
\title{Set seed in R and TensorFlow at the same time}
\usage{
keras_set_seed(seed)
}
\arguments{
\item{seed}{1 integer value.}
}
\description{
Some Keras models requires seeds to be set in both R and TensorFlow to
achieve reproducible results. This function sets these seeds at the same
time using version appropriate functions.
}
\keyword{internal}
