% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lending_club}
\alias{lending_club}
\title{Loan Data}
\source{
Lending Club Statistics https://www.lendingclub.com/info/download-data.action
}
\value{
\item{lending_club}{a data frame}
}
\description{
Loan Data
}
\details{
These data were downloaded from the Lending Club
access site (see below) and are from the first quarter of 2016.
A subset of the rows and variables are included here. The
outcome is in the variable \code{Class} and is either "good" (meaning
that the loan was fully paid back or currently on-time) or "bad"
(charged off, defaulted, of 21-120 days late). A data dictionary
can be found on the source website.
}
\examples{
data(lending_club)
str(lending_club)
}
\keyword{datasets}
