% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{make_call}
\alias{make_call}
\title{Make a parsnip call expression}
\usage{
make_call(fun, ns, args, ...)
}
\arguments{
\item{fun}{A character string of a function name.}

\item{ns}{A character string of a package name.}

\item{args}{A named list of argument values.}
}
\value{
A call.
}
\description{
Make a parsnip call expression
}
\details{
The arguments are spliced into the \code{ns::fun()} call. If they are
missing, null, or a single logical, then are not spliced.
}
\keyword{internal}
