% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_classes}
\alias{make_classes}
\title{Prepend a new class}
\usage{
make_classes(prefix)
}
\arguments{
\item{prefix}{A character string for a class.}
}
\value{
A character vector.
}
\description{
This adds an extra class to a base class of "model_spec".
}
\keyword{internal}
