% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_quantiles.R
\name{matrix_to_quantile_pred}
\alias{matrix_to_quantile_pred}
\title{Reformat quantile predictions}
\usage{
matrix_to_quantile_pred(x, object)
}
\arguments{
\item{x}{A matrix of predictions with rows as samples and columns as quantile
levels.}

\item{object}{A parsnip \code{model_fit} object from a quantile regression model.}
}
\description{
Reformat quantile predictions
}
\keyword{internal}
