% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data.R
\name{maybe_matrix}
\alias{maybe_matrix}
\alias{maybe_data_frame}
\title{Fuzzy conversions}
\usage{
maybe_matrix(x)

maybe_data_frame(x)
}
\arguments{
\item{x}{A data frame, matrix, or sparse matrix.}
}
\value{
A data frame, matrix, or sparse matrix.
}
\description{
These are substitutes for \code{as.matrix()} and \code{as.data.frame()} that leave
a sparse matrix as-is.
}
