% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlp.R
\name{mcqrnn_wrap}
\alias{mcqrnn_wrap}
\title{Wrapper for \code{qrnn::mcqrnn.fit()}}
\usage{
mcqrnn_wrap(x, y, ..., trace, tau)
}
\arguments{
\item{x}{Predictor matrix}

\item{y}{outcome vector}

\item{...}{Options to pass to \code{qrnn::mcqrnn.fit()}}

\item{trace}{Logical for printing}

\item{tau}{Vector of quantile levels.}
}
\description{
Wrapper for \code{qrnn::mcqrnn.fit()}
}
\keyword{internal}
