% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{model_printer}
\alias{model_printer}
\title{Print helper for model objects}
\usage{
model_printer(x, ...)
}
\arguments{
\item{x}{A model object.}

\item{...}{Not currently used.}
}
\description{
A common format function that prints information about the model object (e.g.
arguments, calls, packages, etc).
}
\keyword{internal}
