% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{multi_predict.model_fit}
\alias{multi_predict.model_fit}
\title{TODO}
\usage{
\method{multi_predict}{model_fit}(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\value{
TODO
}
\description{
TODO
}
