% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_reg.R
\name{multinom_reg}
\alias{multinom_reg}
\title{General Interface for Multinomial Regression Models}
\usage{
multinom_reg(mode = "classification", penalty = NULL, mixture = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "classification".}

\item{penalty}{A non-negative number representing the total
amount of regularization (\code{glmnet}, \code{keras}, and \code{spark} only).
For \code{keras} models, this corresponds to purely L2 regularization
(aka weight decay) while the other models can be a combination
of L1 and L2 (depending on the value of \code{mixture}).}

\item{mixture}{A number between zero and one (inclusive) that is the
proportion of L1 regularization (i.e. lasso) in the model. When
\code{mixture = 1}, it is a pure lasso model while \code{mixture = 0} indicates that
ridge regression is being used. (\code{glmnet} and \code{spark} only).}
}
\description{
\code{multinom_reg()} fits a model that uses linear predictors to predict multi-class
data using the multinomial distribution.

There are different ways to fit this model. Information about the available
\emph{engines} that can be used for fitting at:

\Sexpr[stage=render,results=rd]{parsnip:::find_engine_files("multinom_reg")}

More information on how \code{parsnip} is used for model is at
\url{https://www.tidymodels.org}.
}
\details{

}
\examples{
show_engines("multinom_reg")

multinom_reg()
# Parameters can be represented by a placeholder:
multinom_reg(penalty = varying())
}
\seealso{
\code{\link[=fit.model_spec]{fit.model_spec()}}, \code{\link[=set_engine]{set_engine()}}, \code{\link[=update]{update()}}
}
