% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsedms.R
\name{parsedms}
\alias{parsedms}
\title{Parse degrees minutes seconds}
\usage{
parsedms(input)
}
\arguments{
\item{input}{string value}
}
\value{
decimal longitude and latitude
}
\description{
Parse degrees minutes seconds
}

