% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip-package.R
\docType{package}
\name{parsnip-package}
\alias{parsnip}
\alias{parsnip-package}
\title{parsnip}
\description{
The goal of parsnip is to provide a tidy, unified interface to models that
can be used to try a range of models without getting bogged down in the
syntactical minutiae of the underlying packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidymodels/parsnip}
  \item \url{https://parsnip.tidymodels.org/}
  \item Report bugs at \url{https://github.com/tidymodels/parsnip/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Authors:
\itemize{
  \item Davis Vaughan \email{davis@posit.co}
}

Other contributors:
\itemize{
  \item Emil Hvitfeldt \email{emil.hvitfeldt@posit.co} [contributor]
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
