% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip.R
\docType{package}
\name{parsnip}
\alias{parsnip}
\title{Parsing geographical coordinate strings}
\description{
Parsing geographical coordinate strings
}
\author{
Pieter Provoost, \email{pieterprovoost@gmail.com}
}
