% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_in.R
\name{parsnip_addin}
\alias{parsnip_addin}
\title{Start an RStudio Addin that can write model specifications}
\usage{
parsnip_addin()
}
\description{
\code{parsnip_addin()} starts a process in the RStudio IDE Viewer window
that allows users to write code for parsnip model specifications from
various R packages. The new code is written to the current document at the
location of the cursor.
}
