% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict_class.R,
%   R/predict_classprob.R
\name{predict.model_fit}
\alias{predict.model_fit}
\alias{predict_class.model_fit}
\alias{predict_class}
\alias{predict_classprob.model_fit}
\alias{predict_classprob}
\title{Model Predictions}
\usage{
\method{predict}{model_fit}(object, newdata, ...)

\method{predict_class}{model_fit}(object, newdata, ...)

predict_class(object, ...)

\method{predict_classprob}{model_fit}(object, newdata, ...)

predict_classprob(object, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{newdata}{A rectangular data object, such as a data frame.}

\item{...}{Arguments to pass to other methods (not currently used).}
}
\value{
\code{predict} returns an unnamed numeric vector. \code{predict_class} returns
a factor variable and \code{predict_classprob} generates a tibble with a column
for each class level.
}
\description{
Apply a model to create different types of predictions. \code{predict} is used
when the outcome is a simple number. For classification models,
\code{predict_class} and  \code{predict_classprob} can be used.
}
