% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_class.R
\name{predict_class.model_fit}
\alias{predict_class.model_fit}
\alias{predict_class}
\title{Classification Model Predictions}
\usage{
\method{predict_class}{model_fit}(object, newdata, ...)

predict_class(object, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{newdata}{A rectangular data object, such as a data frame.}

\item{...}{Arguments to pass to other methods (not currently used).}
}
\value{
An unnamed factor vector of predictions with the same
levels as the original outcome.
}
\description{
Apply a model to create predictions when the outcome is
qualitative.
}
