% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_classprob.R
\name{predict_classprob.model_fit}
\alias{predict_classprob.model_fit}
\alias{predict_classprob}
\title{Classification Model Probability Predictions}
\usage{
\method{predict_classprob}{model_fit}(object, newdata, ...)

predict_classprob(object, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{newdata}{A rectangular data object, such as a data frame.}

\item{...}{Arguments to pass to other methods (not currently used).}
}
\value{
A tibble with columns for each class.
}
\description{
Apply a model to create class probability predictions when the outcome is
qualitative.
}
