% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_reg.R
\name{predict_raw._fishnet}
\alias{predict_raw._fishnet}
\alias{multi_predict._fishnet}
\title{Model predictions across many sub-models}
\usage{
\method{predict_raw}{`_fishnet`}(object, new_data, opts = list(), ...)

\method{multi_predict}{`_fishnet`}(object, new_data, type = NULL, penalty = NULL, ...)
}
\arguments{
\item{object}{A \code{model_fit} object.}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{opts}{A list of options..}

\item{...}{Optional arguments to pass to \code{predict.model_fit(type = "raw")}
such as \code{type}.}

\item{penalty}{A numeric vector of penalty values.}
}
\value{
A tibble with the same number of rows as the data being predicted.
There is a list-column named \code{.pred} that contains tibbles with
multiple rows per sub-model.
}
\description{
For some models, predictions can be made on sub-models in the model object.
}
\keyword{internal}
