% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data based on parsnip encoding information}
\usage{
prepare_data(object, new_data)
}
\arguments{
\item{object}{A parsnip model object}

\item{new_data}{A data frame}
}
\value{
A data frame or matrix
}
\description{
Prepare data based on parsnip encoding information
}
\keyword{internal}
