% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_quantiles.R
\name{quantile_pred}
\alias{quantile_pred}
\alias{extract_quantile_levels}
\title{Create a vector containing sets of quantiles}
\usage{
quantile_pred(values, quantile_levels = double())

extract_quantile_levels(x)
}
\arguments{
\item{values}{A matrix of values. Each column should correspond to one of
the quantile levels.}

\item{quantile_levels}{A vector of probabilities corresponding to \code{values}.}
}
\value{
A vector of values associated with the quantile levels.
}
\description{
Create a vector containing sets of quantiles
}
\examples{
v <- quantile_pred(matrix(rnorm(20), 5), c(.2, .4, .6, .8))

# Access the underlying information
attr(v, "quantile_levels")
unclass(v)

# tidy format
as_tibble(v)

# matrix format
as.matrix(v)
}
