% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/varying.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{\%>\%}
\alias{fit}
\alias{fit_xy}
\alias{tidy}
\alias{glance}
\alias{augment}
\alias{required_pkgs}
\alias{contr_one_hot}
\alias{extract_spec_parsnip}
\alias{extract_fit_engine}
\alias{extract_parameter_set_dials}
\alias{extract_parameter_dials}
\alias{tune}
\alias{frequency_weights}
\alias{importance_weights}
\alias{extract_fit_time}
\alias{varying_args}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics:augment]{augment()}}, \code{\link[generics:fit]{fit()}}, \code{\link[generics:fit_xy]{fit_xy()}}, \code{\link[generics:glance]{glance()}}, \code{\link[generics:required_pkgs]{required_pkgs()}}, \code{\link[generics:tidy]{tidy()}}, \code{\link[generics:varying_args]{varying_args()}}}

  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}}

  \item{hardhat}{\code{\link[hardhat:contr_one_hot]{contr_one_hot()}}, \code{\link[hardhat:extract_fit_engine]{extract_fit_engine()}}, \code{\link[hardhat:extract_fit_time]{extract_fit_time()}}, \code{\link[hardhat:extract_parameter_dials]{extract_parameter_dials()}}, \code{\link[hardhat:extract_parameter_set_dials]{extract_parameter_set_dials()}}, \code{\link[hardhat:extract_spec_parsnip]{extract_spec_parsnip()}}, \code{\link[hardhat:frequency_weights]{frequency_weights()}}, \code{\link[hardhat:importance_weights]{importance_weights()}}, \code{\link[hardhat:tune]{tune()}}}

  \item{magrittr}{\code{\link[magrittr:\%>\%]{\%>\%}}}
}}

