% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{show_call}
\alias{show_call}
\title{Print the model call}
\usage{
show_call(object)
}
\arguments{
\item{object}{A "model_spec" object.}
}
\value{
A character string.
}
\description{
Print the model call
}
\keyword{internal}
