% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsevctrs.R
\name{sparse_data}
\alias{sparse_data}
\title{Using sparse data with parsnip}
\description{
You can figure out whether a given model engine supports sparse data by
calling \code{get_encoding("name of model")} and looking at the \code{allow_sparse_x}
column.
}
\details{
Using sparse data for model fitting and prediction shouldn't require any
additional configurations. Just pass in a sparse matrix such as dgCMatrix
from the \code{Matrix} package or a sparse tibble from the sparsevctrs package
to the data argument of \code{\link[=fit]{fit()}}, \code{\link[=fit_xy]{fit_xy()}}, and \code{\link[=predict]{predict()}}.

Models that don't support sparse data will try to convert to non-sparse data
with warnings. If conversion isn’t possible, an informative error will be
thrown.
}
