% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{stan_conf_int}
\alias{stan_conf_int}
\title{Wrapper for stan confidence intervals}
\usage{
stan_conf_int(object, newdata)
}
\arguments{
\item{object}{A stan model fit}

\item{newdata}{A data set.}
}
\description{
Wrapper for stan confidence intervals
}
\keyword{internal}
