% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipcw.R
\name{surv-helpers}
\alias{surv-helpers}
\alias{.is_censored_right}
\alias{.check_censored_right}
\alias{.extract_surv_time}
\alias{.extract_surv_status}
\title{Helper functions for Surv objects}
\usage{
.is_censored_right(surv)

.check_censored_right(surv)

.extract_surv_time(surv)

.extract_surv_status(surv)
}
\arguments{
\item{surv}{A \code{\link[survival:Surv]{survival::Surv()}} object}
}
\value{
\itemize{
\item \code{.extract_status()} returns a vector.
\item \code{.extract_time()} returns a vector when the type is \code{"right"} or \code{"left"}
and a tibble otherwise.
\item Functions starting with \code{.is_} or \code{.check_} return logicals although the
latter will fail when \code{FALSE}.
}
}
\description{
Functions to make it easier to work with \code{\link[survival:Surv]{survival::Surv()}} objects.
}
\details{
\code{.extract_status()} will return the data as 0/1 even if the original object
used the legacy encoding of 1/2. See \code{\link[survival:Surv]{survival::Surv()}}.
}
\keyword{internal}
