% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-survival.R
\name{survival-helpers}
\alias{survival-helpers}
\alias{.extract_surv_time}
\title{Helpers for survival analysis}
\usage{
.extract_surv_time(surv)
}
\arguments{
\item{surv}{A single \code{\link[survival:Surv]{survival::Surv()}} object.}
}
\value{
\itemize{
\item \code{.extract_surv_status()} returns a vector.
\item \code{.extract_surv_time()} returns a vector when the type is \code{"right"} or
\code{"left"} and a tibble otherwise.
}
}
\description{
These functions make it a little easier to work with \code{\link[survival:Surv]{survival::Surv()}}
objects.
}
