% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_auto_int.R
\name{tabular_auto_int}
\alias{tabular_auto_int}
\title{AutoInt: Automatic Feature Interaction Learning}
\usage{
tabular_auto_int(
  mode = "unknown",
  engine = "brulee",
  epochs = NULL,
  num_embedding = NULL,
  hidden_units = NULL,
  hidden_activations = NULL,
  num_attn_feat = NULL,
  num_attn_heads = NULL,
  num_attn_blocks = NULL,
  activation = NULL,
  dropout = NULL,
  dropout_attn = NULL,
  dropout_embedding = NULL,
  penalty = NULL,
  mixture = NULL,
  learn_rate = NULL,
  rate_schedule = NULL,
  momentum = NULL,
  batch_size = NULL,
  class_weights = NULL,
  stop_iter = NULL
)
}
\arguments{
\item{mode}{A single character string for the prediction outcome mode.
Possible values for this model are "unknown", "regression", or
"classification".}

\item{engine}{A single character string specifying what computational engine
to use for fitting.}

\item{epochs}{An integer for the number of training iterations.}

\item{num_embedding}{An integer for the dimensionality of the embedding space
for features.}

\item{hidden_units}{An integer vector for the number of units in the hidden
layers after the attention mechanism.}

\item{hidden_activations}{A character vector denoting the activation functions
for the hidden layers.}

\item{num_attn_feat}{An integer for the number of attention features.}

\item{num_attn_heads}{An integer for the number of attention heads in the
multi-head attention mechanism.}

\item{num_attn_blocks}{An integer for the number of sequential attention
blocks.}

\item{activation}{A character vector denoting the type of relationship
between the different layers. The activation function between the hidden and
output layers is automatically set to either "linear" or "softmax"
depending on the type of outcome. Possible values depend on the engine
being used.}

\item{dropout}{A number between 0 (inclusive) and 1 denoting the proportion
of model parameters randomly set to zero during model training.}

\item{dropout_attn}{A number between 0 (inclusive) and 1 denoting the
proportion of attention weights set to zero during model training.}

\item{dropout_embedding}{A number between 0 (inclusive) and 1 denoting the
proportion of embedding values set to zero during model training.}

\item{penalty}{A non-negative numeric value for the amount of weight
decay.}

\item{mixture}{A number between zero and one (inclusive) denoting the
proportion of L1 regularization (i.e. lasso) in the model.
\itemize{
\item \code{mixture = 1} specifies a pure lasso model,
\item \code{mixture = 0}  specifies a ridge regression model, and
\item \verb{0 < mixture < 1} specifies an elastic net model, interpolating lasso and ridge.
}

Available for specific engines only.}

\item{learn_rate}{A number for the rate at which the boosting algorithm adapts
from iteration-to-iteration (specific engines only). This is sometimes referred to
as the shrinkage parameter.}

\item{rate_schedule}{A character string for the learning rate schedule.}

\item{momentum}{A number for the momentum parameter in optimizers that use it.}

\item{batch_size}{An integer for the number of training instances in each
batch.}

\item{class_weights}{Numeric class weights for imbalanced data
(classification only).}

\item{stop_iter}{The number of iterations without improvement before
stopping (specific engines only).}
}
\description{
\code{tabular_auto_int()} uses an attention mechanism to automatically learn
embedding co-representations for tabular data. This function can fit
classification and regression models.

\Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("tabular_auto_int")}

More information on how \pkg{parsnip} is used for modeling is at
\url{https://www.tidymodels.org/}.
}
\examples{
\dontshow{if (!parsnip:::is_cran_check()) withAutoprint(\{ # examplesIf}
show_engines("tabular_auto_int")

tabular_auto_int(mode = "classification", num_attn_blocks = 4)
\dontshow{\}) # examplesIf}
}
\seealso{
\Sexpr[stage=render,results=rd]{parsnip:::make_seealso_list("tabular_auto_int")}, \code{\link[brulee:brulee_auto_int]{brulee::brulee_auto_int()}}
}
