% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_icl.R
\name{tabular_icl}
\alias{tabular_icl}
\title{TabICL: prior data fitted networks}
\usage{
tabular_icl(
  mode = "unknown",
  engine = "brulee",
  num_estimators = NULL,
  softmax_temperature = NULL
)
}
\arguments{
\item{mode}{A single character value for the type of model.
The possible values for this model are "classification" and "regression".}

\item{engine}{A single character string specifying what computational engine
to use for fitting. Possible engines are listed below. The default for this
model is \code{"brulee"}.}

\item{num_estimators}{An integer for the ensemble size. Default is \code{8L}.}

\item{softmax_temperature}{An adjustment factor that is a divisor in the
exponents of the softmax function (see \code{\link[brulee:brulee_tab_icl]{brulee::brulee_tab_icl()}}). Defaults
to 0.9.}
}
\description{
This function uses a pre-trained deep learning network that emulates Bayesian
inference. The model was trained on a large number of simulated data sets
and an attention mechanism is use to make relevant predictions for specific
(i.e., real) data sets.

\Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("tabular_icl")}

More information on how \pkg{parsnip} is used for modeling is at
\url{https://www.tidymodels.org/}.
}
\details{
This function fits classification and regression models.
}
\examples{
\dontshow{if (!parsnip:::is_cran_check()) withAutoprint(\{ # examplesIf}
show_engines("tabular_icl")

tabular_icl()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://github.com/soda-inria/tabicl}

Qu, J., Holzmüller, D., Varoquaux, G., & Morvan, M. L. (2025). Tabicl: A
tabular foundation model for in-context learning on large data. arXiv
preprint arXiv:2502.05564.

Qu, J., Holzmüller, D., Varoquaux, G., & Morvan, M. L. (2026). TabICLv2: A
better, faster, scalable, and open tabular foundation model. arXiv preprint
arXiv:2602.11139.
}
\seealso{
\Sexpr[stage=render,results=rd]{parsnip:::make_seealso_list("tabular_icl")}, \code{\link[brulee:brulee_tab_icl]{brulee::brulee_tab_icl()}}
}
