% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_pfn.R
\name{tabular_pfn}
\alias{tabular_pfn}
\title{TabPFN: prior data fitted networks}
\usage{
tabular_pfn(
  mode = "unknown",
  engine = "tabpfn",
  num_estimators = NULL,
  softmax_temperature = NULL,
  balance_probabilities = NULL,
  average_before_softmax = NULL
)
}
\arguments{
\item{mode}{A single character value for the type of model.
The possible values for this model are "classification" and "regression".}

\item{engine}{A single character string specifying what computational engine
to use for fitting. Possible engines are listed below. The default for this
model is \code{"tabpfn"}.}

\item{num_estimators}{An integer for the ensemble size. Default is \code{8L}.}

\item{softmax_temperature}{An adjustment factor that is a divisor in the
exponents of the softmax function (see \code{\link[tabpfn:tab_pfn]{tabpfn::tab_pfn()}}). Defaults to 0.9.}

\item{balance_probabilities}{A logical to adjust the prior probabilities in
cases where there is a class imbalance. Default is \code{FALSE}. Classification
only.}

\item{average_before_softmax}{A logical. For cases where
\code{num_estimators > 1}, should the average be done before using the softmax
function or after? Default is \code{FALSE}.}
}
\description{
This function uses a pre-trained deep learning network that emulates Bayesian
inference. The model was trained on a large number of simulated data sets
and an attention mechanism is use to make relevant predictions for specific
(i.e., real) data sets.

\Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("tabular_pfn")}

More information on how \pkg{parsnip} is used for modeling is at
\url{https://www.tidymodels.org/}.
}
\details{
This function fits classification and regression models.
}
\examples{
\dontshow{if (!parsnip:::is_cran_check()) withAutoprint(\{ # examplesIf}
show_engines("tabular_pfn")

tabular_pfn()
\dontshow{\}) # examplesIf}
}
\references{
\url{https://github.com/PriorLabs/TabPFN}

Hollmann, Noah, Samuel Müller, Lennart Purucker, Arjun Krishnakumar, Max
Körfer, Shi Bin Hoo, Robin Tibor Schirrmeister, and Frank Hutter.
"Accurate predictions on small data with a tabular foundation model."
\emph{Nature} 637, no. 8045 (2025): 319-326.

Hollmann, Noah, Samuel Müller, Katharina Eggensperger, and Frank Hutter.
"Tabpfn: A transformer that solves small tabular classification problems in
a second." \emph{arXiv preprint} arXiv:2207.01848 (2022).

Müller, Samuel, Noah Hollmann, Sebastian Pineda Arango, Josif Grabocka, and
Frank Hutter. "Transformers can do Bayesian inference." \emph{arXiv preprint}
arXiv:2112.10510 (2021).
}
\seealso{
\Sexpr[stage=render,results=rd]{parsnip:::make_seealso_list("tabular_pfn")}, \code{\link[tabpfn:tab_pfn]{tabpfn::tab_pfn()}}
}
