% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_resnet.R
\name{tabular_resnet}
\alias{tabular_resnet}
\title{Residual Neural Network for Tabular Data}
\usage{
tabular_resnet(
  mode = "unknown",
  engine = "brulee",
  hidden_units = NULL,
  bottleneck_units = NULL,
  residual_at = NULL,
  penalty = NULL,
  mixture = NULL,
  dropout = NULL,
  epochs = NULL,
  activation = NULL,
  learn_rate = NULL,
  rate_schedule = NULL,
  momentum = NULL,
  batch_size = NULL,
  class_weights = NULL,
  stop_iter = NULL
)
}
\arguments{
\item{mode}{A single character string for the prediction outcome mode.
Possible values for this model are "unknown", "regression", or
"classification".}

\item{engine}{A single character string specifying what computational engine
to use for fitting.}

\item{hidden_units}{An integer vector for the number of units in the hidden
model.}

\item{bottleneck_units}{The number of embeddings that are produced by batch
normalization.}

\item{residual_at}{An integer vector with the layer number should use a
residual connection (i.e., skip layer).}

\item{penalty}{A non-negative numeric value for the amount of weight
decay.}

\item{mixture}{A number between zero and one (inclusive) giving the
proportion of L1 regularization (i.e. lasso) in the model. \code{mixture = 1}
is a pure lasso model while \code{mixture = 0} indicates ridge regression
(a.k.a weight decay).}

\item{dropout}{A number between 0 (inclusive) and 1 denoting the proportion
of model parameters randomly set to zero during model training.}

\item{epochs}{An integer for the number of training iterations.}

\item{activation}{A vector character strings denoting the type of relationship
between the layers. The activation
function between the hidden and output layers is automatically set to either
"linear" or "softmax" depending on the type of outcome. Possible values
depend on the engine being used.}

\item{learn_rate}{A number for the rate at which the boosting algorithm adapts
from iteration-to-iteration (specific engines only). This is sometimes referred to
as the shrinkage parameter.}

\item{rate_schedule}{A character string for the learning rate schedule.}

\item{momentum}{A number for the momentum parameter in optimizers that use it.}

\item{batch_size}{An integer for the number of training instances in each
batch.}

\item{class_weights}{Numeric class weights for imbalanced data
(classification only).}

\item{stop_iter}{The number of iterations without improvement before
stopping (specific engines only).}
}
\description{
\code{tabular_resnet()} ... This function can fit classification and
regression models.

\Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("tabular_resnet")}

More information on how \pkg{parsnip} is used for modeling is at
\url{https://www.tidymodels.org/}.
}
\examples{
\dontshow{if (!parsnip:::is_cran_check()) withAutoprint(\{ # examplesIf}
show_engines("tabular_resnet")

tabular_resnet(mode = "classification", penalty = 0.01)
\dontshow{\}) # examplesIf}
}
\seealso{
\Sexpr[stage=render,results=rd]{parsnip:::make_seealso_list("tabular_resnet")}, \code{\link[brulee:brulee_resnet]{brulee::brulee_resnet()}}
}
