% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_glmnet.R
\name{tidy._elnet}
\alias{tidy._elnet}
\alias{tidy._lognet}
\alias{tidy._multnet}
\alias{tidy._fishnet}
\alias{tidy._coxnet}
\title{tidy methods for glmnet models}
\usage{
\method{tidy}{`_elnet`}(x, penalty = NULL, ...)

\method{tidy}{`_lognet`}(x, penalty = NULL, ...)

\method{tidy}{`_multnet`}(x, penalty = NULL, ...)

\method{tidy}{`_fishnet`}(x, penalty = NULL, ...)

\method{tidy}{`_coxnet`}(x, penalty = NULL, ...)
}
\arguments{
\item{x}{A fitted parsnip model that used the \code{glmnet} engine.}

\item{penalty}{A \emph{single} numeric value. If none is given, the value specified
in the model specification is used.}

\item{...}{Not used}
}
\value{
A tibble with columns \code{term}, \code{estimate}, and \code{penalty}. When a
multinomial mode is used, an additional \code{class} column is included.
}
\description{
\code{tidy()} methods for the various \code{glmnet} models that return the coefficients
for the specific penalty value used by the parsnip model fit.
}
\keyword{internal}
