% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_liblinear.R
\name{tidy._LiblineaR}
\alias{tidy._LiblineaR}
\title{tidy methods for LiblineaR models}
\usage{
\method{tidy}{`_LiblineaR`}(x, ...)
}
\arguments{
\item{x}{A fitted parsnip model that used the \code{LiblineaR} engine.}

\item{...}{Not used}
}
\value{
A tibble with columns \code{term} and \code{estimate}.
}
\description{
\code{tidy()} methods for the various \code{LiblineaR} models that return the
coefficients from the parsnip model fit.
}
\keyword{internal}
