% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Resolve a Model Specification for a Computational Engine}
\usage{
translate(x, ...)
}
\arguments{
\item{x}{A model specification.}

\item{...}{Not currently used.}
}
\description{
\code{translate} will translate a model specification into a code
object that is specific to a particular engine (e.g. R package).
It translates generic parameters to their counterparts.
}
