% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C5_rules.R, R/cubist_rules.R, R/rule_fit.R
\name{tunable.C5_rules}
\alias{tunable.C5_rules}
\alias{tunable-parsnip}
\alias{tunable.cubist_rules}
\alias{tunable.rule_fit}
\title{Tunable methods}
\usage{
tunable.C5_rules(x, ...)

tunable.cubist_rules(x, ...)

tunable.rule_fit(x, ...)
}
\description{
Tunable methods
}
\keyword{internal}
