% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_sum.R
\name{type_sum.model_spec}
\alias{type_sum.model_spec}
\alias{type_sum.model_fit}
\title{Succinct summary of parsnip object}
\usage{
\method{type_sum}{model_spec}(x)

\method{type_sum}{model_fit}(x)
}
\arguments{
\item{x}{A \code{model_spec} or \code{model_fit} object to summarise.}
}
\value{
A character value.
}
\description{
\code{type_sum} controls how objects are shown when inside tibble
columns.
}
\details{
For \code{model_spec} objects, the summary is "\code{spec[?]}"
or "\code{spec[+]}". The former indicates that either the model
mode has not been declared or that the specification has
\code{varying()} parameters. Otherwise, the latter is shown.

For fitted models, either "\code{fit[x]}" or "\code{fit[+]}" are used
where the "x" implies that the model fit failed in some way.
}
\keyword{internal}
