% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{update.logistic_reg}
\alias{update.logistic_reg}
\title{Update a Logistic Regression Specification}
\usage{
\method{update}{logistic_reg}(object, regularization = NULL, mixture = NULL,
  others = list(), fresh = FALSE, ...)
}
\arguments{
\item{object}{A logistic reression model specification.}

\item{regularization}{An non-negative number representing the
total amount of regularization.}

\item{mixture}{A number between zero and one (inclusive) that
represents the proportion of regularization that is used for the
L2 penalty (i.e. weight decay, or ridge regression) versus L1
(the lasso).}

\item{others}{A named list of arguments to be used by the
underlying models (e.g., \code{stats::glm},
\code{rstanarm::stan_glm}, etc.). These are not evaluated
until the model is fit and will be substituted into the model
fit expression.}

\item{fresh}{A logical for whether the arguments should be
modifed in-place of or replaced wholesale.}

\item{...}{Used for S3 method consistency. Any arguments passed to
the ellipses will result in an error. Use \code{others} instead.}
}
\value{
An updated model specification.
}
\description{
If parameters need to be modified, this function can be used
in lieu of recreating the object from scratch.
}
\examples{
model <- logistic_reg(regularization = 10, mixture = 0.1)
model

update(model, regularization = 1)

update(model, regularization = 1, fresh = TRUE)
}
