% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_forest.R
\name{update.rand_forest}
\alias{update.rand_forest}
\title{Update a Random Forest Specification}
\usage{
\method{update}{rand_forest}(object, mtry = NULL, trees = NULL,
  min_n = NULL, engine_args = list(), fresh = FALSE, ...)
}
\arguments{
\item{object}{A random forest model specification.}

\item{mtry}{An integer for the number of predictors that will
be randomly sampled at each split when creating the tree models.}

\item{trees}{An integer for the number of trees contained in
the ensemble.}

\item{min_n}{An integer for the minimum number of data points
in a node that are required for the node to be split further.}

\item{engine_args}{A named list of arguments to be used by the
underlying models (e.g., \code{ranger::ranger},
\code{randomForest::randomForest}, etc.). These are not evaluated
until the model is fit and will be substituted into the model
fit expression.}

\item{fresh}{A logical for whether the arguments should be
modifed in-place of or replaced wholesale.}

\item{...}{Used for method consistency. Any arguments passed to
the ellipses will result in an error. Use \code{engine_args} instead.}
}
\value{
An updated model specification.
}
\description{
If parameters need to be modified, this function can be used
in lieu of recreating the object from scratch.
}
\examples{
model <- rand_forest(mtry = 10, min_n = 3)
model

update(model, mtry = 1)

update(model, mtry = 1, fresh = TRUE)
}
