% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R, R/parameters.R
\name{varying}
\alias{varying}
\alias{varying}
\title{A Placeholder Function for Argument Values}
\usage{
varying()

varying()
}
\description{
\code{\link[=varying]{varying()}} is used when a parameter will be specified at a later date.

\code{\link[=varying]{varying()}} is used when a parameter will be specified at a later date.
}
