% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_quantiles.R
\name{vec_quantiles}
\alias{vec_quantiles}
\title{Create a vector containing sets of quantiles}
\usage{
vec_quantiles(values, quantile_levels = double())
}
\arguments{
\item{values}{A matrix of values. Each column should correspond to one of
the quantile levels.}

\item{quantile_levels}{A vector of probabilities corresponding to \code{values}.}
}
\value{
A vector of values associated with the quantile levels.
}
\description{
Create a vector containing sets of quantiles
}
\examples{
v <- vec_quantiles(matrix(rnorm(20), 5), c(.2, .4, .6, .8))

# Access the underlying information
attr(v, "quantile_levels")
vctrs::vec_data(v)
}
