% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wa_churn}
\alias{wa_churn}
\title{Watson Churn Data}
\source{
IBM Watson Analytics https://ibm.co/2sOvyvy
}
\value{
\item{wa_churn}{a data frame}
}
\description{
Watson Churn Data
}
\details{
These data were downloaded from the IBM Watson site
(see below) in September 2018. The data contain a factor for
whether a customer churned or not. Alternatively, the \code{tenure}
column presumably contains information on how long the customer
has had an account. A survival analysis can be done on this
column using the \code{churn} outcome as the censoring information. A
data dictionary can be found on the source website.
}
\examples{
data(wa_churn)
str(wa_churn)
}
\keyword{datasets}
