% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_in.R
\name{write_parsnip_specs}
\alias{write_parsnip_specs}
\title{Start an RStudio Addin that can write model specifications}
\usage{
write_parsnip_specs()
}
\description{
\code{write_parsnip_specs()} starts a process in the RStudio IDE Viewer window
that allows users to write code for \code{parsnip} model specifications from
various R packages. The new code are written to the current document at the
location of the cursor.
}
