% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_tree.R
\name{xgb_train}
\alias{xgb_train}
\title{Training helper for xgboost}
\usage{
xgb_train(x, y, max_depth = 6, nrounds = 15, eta = 0.3,
  colsample_bytree = 1, min_child_weight = 1, gamma = 0,
  subsample = 1, ...)
}
\description{
Training helper for xgboost
}
