% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R, R/misc.R, R/varying.R
\name{null_value}
\alias{null_value}
\alias{show_fit}
\alias{update_dot_check}
\alias{new_model_spec}
\alias{check_final_param}
\alias{update_main_parameters}
\alias{update_engine_parameters}
\alias{is_varying}
\title{Functions required for parsnip-adjacent packages}
\usage{
null_value(x)

show_fit(model, eng)

update_dot_check(dots)

new_model_spec(cls, args, eng_args, mode, method, engine)

check_final_param(x)

update_main_parameters(args, param)

update_engine_parameters(eng_args, ...)

is_varying(x)
}
\description{
These functions are helpful when creating new packages that will register
new model specifications.
}
\keyword{internal}
