% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart.R
\name{bart-internal}
\alias{bart-internal}
\alias{bartMachine_interval_calc}
\alias{dbart_predict_calc}
\title{Developer functions for predictions via BART models}
\usage{
bartMachine_interval_calc(new_data, obj, ci = TRUE, level = 0.95)

dbart_predict_calc(obj, new_data, type, level = 0.95, std_err = FALSE)
}
\arguments{
\item{new_data}{A rectangular data object, such as a data frame.}

\item{obj}{A parsnip object.}

\item{ci}{Confidence (TRUE) or prediction interval (FALSE)}

\item{level}{Confidence level.}

\item{type}{A single character value or \code{NULL}. Possible values
are \code{"numeric"}, \code{"class"}, \code{"prob"}, \code{"conf_int"}, \code{"pred_int"},
\code{"quantile"}, \code{"time"}, \code{"hazard"}, \code{"survival"}, or \code{"raw"}. When \code{NULL},
\code{predict()} will choose an appropriate value based on the model's mode.}

\item{std_err}{Attach column for standard error of prediction or not.}
}
\description{
Developer functions for predictions via BART models
}
\keyword{internal}
