% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_models.R
\name{check_mod_val}
\alias{check_mod_val}
\alias{check_mode_val}
\alias{check_engine_val}
\alias{check_arg_val}
\alias{check_submodels_val}
\alias{check_func_val}
\alias{check_fit_info}
\alias{check_pred_info}
\alias{check_pkg_val}
\title{Tools to Check Model Elements}
\usage{
check_mod_val(model, new = FALSE, existence = FALSE)

check_mode_val(mode)

check_engine_val(eng)

check_arg_val(arg)

check_submodels_val(x)

check_func_val(func)

check_fit_info(x)

check_pred_info(x, type)

check_pkg_val(x)
}
\arguments{
\item{model}{A single character string for the model type (e.g.
\code{"rand_forest"}, etc).}

\item{new}{A single logical to check to see if the model that you are check
has not already been registered.}

\item{existence}{A single logical to check to see if the model has already
been registered.}

\item{mode}{A single character string for the model mode (e.g. "regression").}

\item{eng}{A single character string for the model engine.}

\item{arg}{A single character string for the model argument name.}
}
\description{
These functions are similar to constructors and can be used to validate
that there are no conflicts with the underlying model structures used by the
package.
}
\keyword{internal}
