% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{convert_args}
\alias{convert_args}
\title{Make a table of arguments}
\usage{
convert_args(model_name)
}
\arguments{
\item{model_name}{A character string for the model}
}
\description{
Make a table of arguments
}
\keyword{internal}
