% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitter.R
\name{descriptors}
\alias{descriptors}
\title{Data Set Characteristics Available when Fitting Models}
\description{
When using the \code{fit} functions there are some
variables that will be available for use in arguments. For
example, if the user would like to choose an argument value
based on the current number of rows in a data set, the \code{n_obs}
variable can be used. See Details below.
}
\details{
Existing variables:
\itemize{
\item \code{n_obs}: the current number of rows in the data set.
\item \code{n_cols}: the number of columns in the data set that are
associated with the predictors prior to dummy variable creation.
\item \code{n_preds}: the number of predictors after dummy variables
are created (if any).
\item \code{n_facts}: the number of factor predictors in the dat set.
\item \code{n_levs}: If the outcome is a factor, this is a table
with the counts for each level (and \code{NA} otherwise)
}
}
