% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_C5.0.R
\name{details_decision_tree_C5.0}
\alias{details_decision_tree_C5.0}
\title{Decision trees via C5.0}
\description{
\code{\link[C50:C5.0]{C50::C5.0()}} creates a model that is defined as a set of \verb{if/then}
statements that creates a tree-based structure.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 1 tuning parameters:
\itemize{
\item \code{min_n}: Minimal Node Size (type: integer, default: 2L)
}
}

\subsection{Translation from parsnip to the original package (classification)}{\if{html}{\out{<div class="r">}}\preformatted{decision_tree(min_n = integer()) \%>\% 
  set_engine("C5.0") \%>\% 
  set_mode("classification") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Decision Tree Model Specification (classification)
## 
## Main Arguments:
##   min_n = integer()
## 
## Computational engine: C5.0 
## 
## Model fit template:
## parsnip::C5.0_train(x = missing_arg(), y = missing_arg(), weights = missing_arg(), 
##     minCases = integer(), trials = 1)
}

\code{\link[=C5.0_train]{C5.0_train()}} is a wrapper around
\code{\link[C50:C5.0]{C50::C5.0()}} that makes it easier to run this model.
}

\subsection{Preprocessing requirements}{

This engine does not require any special encoding of the predictors.
Categorical predictors can be split into groups without creating dummy
variables for each category.
}

\subsection{References}{
\itemize{
\item Kuhn, M, and K Johnson. 2013. \emph{Applied Predictive Modeling}.
Springer.
}
}
}
\keyword{internal}
