% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_tree_rpartScore.R
\name{details_decision_tree_rpartScore}
\alias{details_decision_tree_rpartScore}
\title{Ordinal decision trees via CART}
\description{
\code{\link[rpartScore:rpartScore]{rpartScore::rpartScore()}} extends \link[=details_decision_tree_rpart]{rpart} to
fit classification trees for ordinal responses.
}
\keyword{internal}
