% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_reg_h2o.R
\name{details_linear_reg_h2o}
\alias{details_linear_reg_h2o}
\title{Linear regression via h2o}
\description{
This model uses regularized least squares to fit models with numeric outcomes.
}
\details{
For this engine, there is a single mode: regression
\subsection{Tuning Parameters}{

This model has 2 tuning parameters:
\itemize{
\item \code{mixture}: Proportion of Lasso Penalty (type: double, default: see
below)
\item \code{penalty}: Amount of Regularization (type: double, default: 0)
}

By default h2o applies no regularization and \code{penalty} is set to zero.
When the engine parameter \code{solver} is set to \code{'L-BFGS'}, \code{mixture}
defaults to 0 (ridge regression) and 0.5 otherwise.
}

\subsection{Translation from parsnip to the original package}{

\code{\link[agua:h2o_train]{agua::h2o_train_glm()}} is a wrapper around
\code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} with \code{family = "gaussian"}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{linear_reg(penalty = 1, mixture = 0.5) \%>\% 
  set_engine("h2o") \%>\% 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Linear Regression Model Specification (regression)
## 
## Main Arguments:
##   penalty = 1
##   mixture = 0.5
## 
## Computational engine: h2o 
## 
## Model fit template:
## agua::h2o_train_glm(x = missing_arg(), y = missing_arg(), lambda = 1, 
##     alpha = 0.5, family = "gaussian")
}\if{html}{\out{</div>}}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine. When using the
formula method via \code{\link[=fit.model_spec]{fit()}}, parsnip
will convert factor columns to indicators.

Predictors should have the same scale. One way to achieve this is to
center and scale each so that each predictor has mean zero and a
variance of one. By default, \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} uses the
argument \code{standardize = TRUE} to center and scale the data.
}
}
\keyword{internal}
