% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom_reg_h2o.R
\name{details_multinom_reg_h2o}
\alias{details_multinom_reg_h2o}
\title{Multinomial regression via h2o}
\description{
\code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} fits a model that uses linear predictors to predict
multiclass data for multinomial responses.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 2 tuning parameters:
\itemize{
\item \code{mixture}: Proportion of Lasso Penalty (type: double, default: see
below)
\item \code{penalty}: Amount of Regularization (type: double, default: 0)
}

By default \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} applies no regularization
and \code{penalty} is set to zero. When the engine parameter \code{solver} is set
to \code{'L-BFGS'}, \code{mixture} defaults to 0 (ridge regression) and 0.5
otherwise.
}

\subsection{Translation from parsnip to the original package}{

\code{\link[agua:h2o_train]{agua::h2o_train_glm()}} for \code{multinom_reg()} is
a wrapper around \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} with
\code{family = 'multinomial'}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{multinom_reg(penalty = double(1), mixture = double(1)) \%>\% 
  set_engine("h2o") \%>\% 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Multinomial Regression Model Specification (classification)
## 
## Main Arguments:
##   penalty = double(1)
##   mixture = double(1)
## 
## Computational engine: h2o 
## 
## Model fit template:
## agua::h2o_train_glm(x = missing_arg(), y = missing_arg(), family = missing_arg(), 
##     lambda = double(1), alpha = double(1), family = "multinomial")
}\if{html}{\out{</div>}}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine. When using the
formula method via \code{\link[=fit.model_spec]{fit()}}, parsnip
will convert factor columns to indicators.

Predictors should have the same scale. One way to achieve this is to
center and scale each so that each predictor has mean zero and a
variance of one.

By default, \code{\link[h2o:h2o.glm]{h2o::h2o.glm()}} uses the argument
\code{standardize = TRUE} to center and scale the data.
}
}
\keyword{internal}
