% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_reg_ordinalNet.R
\name{details_ordinal_reg_ordinalNet}
\alias{details_ordinal_reg_ordinalNet}
\title{Ordinal regression via ordinalNet}
\description{
\code{\link[ordinalNet:ordinalNet]{ordinalNet::ordinalNet()}} uses an original coordinate descent algorithm to
fit models of the elementwise link multinomial-ordinal (ELMO) class, each
comprising parallel and non-parallel forms and including the most common
ordinal regression models, with an elastic net penalty.
}
\keyword{internal}
