% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_reg_vglm.R
\name{details_ordinal_reg_vglm}
\alias{details_ordinal_reg_vglm}
\title{Ordinal regression via vector GLMs}
\description{
\code{\link[VGAM:vglm]{VGAM::vglm()}} fits vector generalized linear models, which specialize to
several families of ordinal regression models.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 2 tuning parameters:
\itemize{
\item \code{ordinal_link}: Ordinal Link (type: character, default: logit)
\item \code{odds_link}: Odds Link (type: character, default: cumulative)
}
}

\subsection{Translation from parsnip to the original package}{

The \strong{ordered} extension package is required to fit this model.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(ordered)

ordinal_reg(ordinal_link = character(0), odds_link = character(0)) \%>\% 
  set_engine("vglm") \%>\% 
  set_mode("classification") \%>\% 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Ordinal Regression Model Specification (classification)
## 
## Main Arguments:
##   ordinal_link = character(0)
##   odds_link = character(0)
## 
## Computational engine: vglm 
## 
## Model fit template:
## ordered::VGAM_vglm_wrapper(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg(), link = character(0), family = character(0), 
##     parallel = TRUE)
}\if{html}{\out{</div>}}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine. When using the
formula method via \code{\link[=fit.model_spec]{fit()}}, parsnip will
convert factor columns to indicators.
}

\subsection{Case weights}{

This model can utilize case weights during model fitting. To use them,
see the documentation in \link{case_weights} and the examples
on \code{tidymodels.org}.

The \code{fit()} and \code{fit_xy()} arguments have arguments called
\code{case_weights} that expect vectors of case weights.

\emph{However}, the documentation in \code{\link[VGAM:vglm]{VGAM::vglm()}} notes
that matrix of case weights can be passed so that different classes have
different weights. tidymodels assumes vector of a weights; a matrix
cannot be passed in.

Also, the engine documentation notes that: “The values of weights must
be positive; try setting a very small value such as 1.0e-8 to
effectively delete an observation.”
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("ordinal_reg_predict") |>
  dplyr::filter(engine == "vglm") |>
  dplyr::select(mode, type)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 2 x 2
##   mode           type 
##   <chr>          <chr>
## 1 classification class
## 2 classification prob
}\if{html}{\out{</div>}}
}

\subsection{References}{
\itemize{
\item Yee T. 2010. The VGAM Package for Categorical Data Analysis. \emph{Journal
of Statistical Software} 32(10):1–34. .
}
}
}
\keyword{internal}
