% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_reg_vglm.R
\name{details_ordinal_reg_vglm}
\alias{details_ordinal_reg_vglm}
\title{Ordinal regression via vector GLMs}
\description{
\code{\link[VGAM:vglm]{VGAM::vglm()}} fits vector generalized linear models, which specialize to
several families of ordinal regression models.
}
\keyword{internal}
