% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_reg_glm.R
\name{details_poisson_reg_glm}
\alias{details_poisson_reg_glm}
\title{Poisson regression via glm}
\description{
\code{\link[stats:glm]{stats::glm()}} uses maximum likelihood to fit a model for count data.
}
\details{
For this engine, there is a single mode: regression
\subsection{Tuning Parameters}{

This engine has no tuning parameters.
}

\subsection{Translation from parsnip to the underlying model call (regression)}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{poisson_reg() \%>\%
  set_engine("glm") \%>\%
  translate()
}\if{html}{\out{</div>}}\preformatted{## Poisson Regression Model Specification (regression)
## 
## Computational engine: glm 
## 
## Model fit template:
## stats::glm(formula = missing_arg(), data = missing_arg(), weights = missing_arg(), 
##     family = stats::poisson)
}
}

\subsection{Preprocessing requirements}{

Factor/categorical predictors need to be converted to numeric values
(e.g., dummy or indicator variables) for this engine. When using the
formula method via
\code{\link[=fit.model_spec]{fit.model_spec()}}, parsnip will
convert factor columns to indicators.
}
}
\keyword{internal}
