% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_forest_ordinalForest.R
\name{details_rand_forest_ordinalForest}
\alias{details_rand_forest_ordinalForest}
\title{Random forests via ordinalForest}
\description{
\code{\link[ordinalForest:ordfor]{ordinalForest::ordfor()}} fits a model by creating a large number of
regression forests using different scorings of an ordinal response, then
creating a single regression forest based on an optimal subset of scorings.
}
\keyword{internal}
