% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_forest_ordinalForest.R
\name{details_rand_forest_ordinalForest}
\alias{details_rand_forest_ordinalForest}
\title{Random forests via ordinalForest}
\description{
\code{\link[ordinalForest:ordfor]{ordinalForest::ordfor()}} fits a model by creating a large number of
regression forests using different scorings of an ordinal response, then
creating a single regression forest based on an optimal subset of scorings.
}
\details{
For this engine, there is a single mode: classification
\subsection{Tuning Parameters}{

This model has 3 tuning parameters:
\itemize{
\item \code{mtry}: # Randomly Selected Predictors (type: integer, default: see
below)
\item \code{trees}: # Trees (type: integer, default: 500L)
\item \code{min_n}: Minimal Node Size (type: integer, default: see below)
}

\code{mtry} depends on the number of columns and the model mode. The default
in \code{\link[ordinalForest:ordfor]{ordinalForest::ordfor()}} is
\code{floor(sqrt(ncol(x)))}.

\code{min_n} depends on the mode. For regression, a value of 5 is the
default. For classification, a value of 10 is used.
}

\subsection{Translation from parsnip to the original package (classification)}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{rand_forest(
  mtry = integer(1),
  trees = integer(1),
  min_n = integer(1)
) |> 
  set_engine("ordinalForest") |> 
  set_mode("classification") |> 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Random Forest Model Specification (classification)
## 
## Main Arguments:
##   mtry = integer(1)
##   trees = integer(1)
##   min_n = integer(1)
## 
## Computational engine: ordinalForest 
## 
## Model fit template:
## ordered::ordinalForest_wrapper(x = missing_arg(), y = missing_arg(), 
##     mtry = min_cols(~integer(1), x), ntreefinal = integer(1), 
##     min.node.size = min_rows(~integer(1), x), num.threads = 1, 
##     perffunction = "probability")
}\if{html}{\out{</div>}}
}

\subsection{Preprocessing requirements}{

This engine does not require any special encoding of the predictors.
Categorical predictors can be partitioned into groups of factor levels
(e.g. \verb{\{a, c\}} vs \verb{\{b, d\}}) when splitting at a node. Dummy variables
are not required for this model.
}

\subsection{Case weights}{

This model can utilize case weights during model fitting. To use them,
see the documentation in \link{case_weights} and the examples
on \code{tidymodels.org}.

The \code{fit()} and \code{fit_xy()} functions have arguments called
\code{case_weights} that expect vectors of case weights.
}

\subsection{Other notes}{

By default, parallel processing is turned off. When tuning, it is more
efficient to parallelize over the resamples and tuning parameters. To
parallelize the construction of the trees within the model, change the
\code{num.threads} argument via \code{\link[=set_engine]{set_engine()}}.
}

\subsection{References}{
\itemize{
\item Hornung R. 2020. Ordinal Forests. \emph{Journal of Classification} 37:4-17.
.
\item Kuhn, M, and K Johnson. 2013. \emph{Applied Predictive Modeling}. Springer.
}
}
}
\keyword{internal}
