% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_forest_randomForestSRC.R
\name{details_rand_forest_randomForestSRC}
\alias{details_rand_forest_randomForestSRC}
\title{Random survival forests via randomForestSRC}
\description{
\code{\link[randomForestSRC:rfsrc]{randomForestSRC::rfsrc()}} fits a random survival forest: a large number of
survival trees, each grown on a bootstrap sample of the data. The final
prediction averages the predictions from the individual trees.
}
\details{
For this engine, there is a single mode: censored regression
\subsection{Tuning Parameters}{

This model has 3 tuning parameters:
\itemize{
\item \code{trees}: # Trees (type: integer, default: 500L)
\item \code{min_n}: Minimal Node Size (type: integer, default: 15L)
\item \code{mtry}: # Randomly Selected Predictors (type: integer, default:
ceiling(sqrt(n_predictors)))
}
}

\subsection{Translation from parsnip to the original package (censored regression)}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(censored)

rand_forest() |>
  set_engine("randomForestSRC") |>
  set_mode("censored regression") |>
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Random Forest Model Specification (censored regression)
## 
## Computational engine: randomForestSRC 
## 
## Model fit template:
## censored::rfsrc_train(formula = missing_arg(), data = missing_arg(), 
##     weights = missing_arg())
}\if{html}{\out{</div>}}

\code{censored::rfsrc_train()} is a wrapper around
\code{\link[randomForestSRC:rfsrc]{randomForestSRC::rfsrc()}} that makes it
easier to run this model.
}

\subsection{Preprocessing requirements}{

This engine does not require any special encoding of the predictors.
Categorical predictors can be partitioned into groups of factor levels
(e.g. \verb{\{a, c\}} vs \verb{\{b, d\}}) when splitting at a node. Dummy variables
are not required for this model.
}

\subsection{Case weights}{

This model can utilize case weights during model fitting. To use them,
see the documentation in \link{case_weights} and the examples
on \code{tidymodels.org}.

The \code{fit()} and \code{fit_xy()} functions have arguments called
\code{case_weights} that expect vectors of case weights.
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("rand_forest_predict") |>
  dplyr::filter(engine == "randomForestSRC") |>
  dplyr::select(mode, type) |>
  print(n = Inf)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 2 x 2
##   mode                type    
##   <chr>               <chr>   
## 1 censored regression time    
## 2 censored regression survival
}\if{html}{\out{</div>}}
}

\subsection{Other details}{

Predictions of type \code{"time"} are predictions of the median survival
time.
}

\subsection{References}{
\itemize{
\item Ishwaran H, Kogalur UB, Blackstone EH, Lauer MS. Random survival
forests. Annals of Applied Statistics 2008; 2(3):841-860. .
}
}
}
\keyword{internal}
