% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabular_rln_brulee.R
\name{details_tabular_rln_brulee}
\alias{details_tabular_rln_brulee}
\title{Regularization learning network via brulee}
\description{
\code{\link[brulee:brulee_rln]{brulee::brulee_rln()}} fits a regularization learning network.
}
\details{
For this engine, there is a single mode: regression
\subsection{Tuning Parameters}{

This model has 11 tuning parameters:
\itemize{
\item \code{hidden_units}: # Hidden Units (type: integer, default: 5L)
\item \code{penalty_type}: Penalty Type (type: character, default: ‘L1’)
\item \code{penalty_average}: Penalty Average (type: double, default: 1e-10)
\item \code{step_rate}: Step Rate (type: double, default: 1e6)
\item \code{activation}: Activation Function (type: character, default: ‘relu’)
\item \code{epochs}: # Epochs (type: integer, default: 100L)
\item \code{learn_rate}: Learning Rate (type: double, default: 0.001)
\item \code{rate_schedule}: Learning Rate Scheduler (type: character, default:
‘none’)
\item \code{momentum}: Gradient Descent Momentum (type: double, default: 0.0)
\item \code{batch_size}: Batch Size (type: integer, default: NULL)
\item \code{stop_iter}: # Iterations Before Stopping (type: integer, default:
20L)
}

\code{penalty_average} and \code{step_rate} are specified on the natural scale but
are best tuned on the log10 scale.
}

\subsection{Translation from parsnip to the original package (regression)}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{tabular_rln(
  hidden_units = integer(1),
  penalty_type = character(1),
  penalty_average = double(1),
  step_rate = double(1),
  activation = character(1),
  epochs = integer(1),
  learn_rate = double(1),
  rate_schedule = character(1),
  momentum = double(1),
  batch_size = NULL,
  stop_iter = integer(1)
) |>
  set_engine("brulee") |>
  set_mode("regression") |>
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## tabular rln Model Specification (regression)
## 
## Main Arguments:
##   hidden_units = integer(1)
##   penalty_type = character(1)
##   penalty_average = double(1)
##   step_rate = double(1)
##   activation = character(1)
##   epochs = integer(1)
##   learn_rate = double(1)
##   rate_schedule = character(1)
##   momentum = double(1)
##   stop_iter = integer(1)
## 
## Computational engine: brulee 
## 
## Model fit template:
## brulee::brulee_rln(x = missing_arg(), y = missing_arg(), hidden_units = integer(1), 
##     penalty_type = character(1), penalty_average = double(1), 
##     step_rate = double(1), activation = character(1), epochs = integer(1), 
##     learn_rate = double(1), rate_schedule = character(1), momentum = double(1), 
##     stop_iter = integer(1))
}\if{html}{\out{</div>}}
}

\subsection{Preprocessing requirements}{

\code{brulee_rln()} requires numeric predictors. Factor or categorical
predictors need to be converted to numeric values (e.g., dummy or
indicator variables) before fitting; parsnip does \emph{not} create indicator
variables for this engine, so use a recipe (or some other method) to
make them numeric.

Predictors should have the same scale. One way to achieve this is to
center and scale each so that each predictor has mean zero and a
variance of one.
}

\subsection{Case weights}{

The underlying model implementation does not allow for case weights.
}

\subsection{Prediction types}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{parsnip:::get_from_env("tabular_rln_predict") |>
  dplyr::filter(engine == "brulee") |>
  dplyr::select(mode, type)
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## # A tibble: 1 x 2
##   mode       type   
##   <chr>      <chr>  
## 1 regression numeric
}\if{html}{\out{</div>}}
}

\subsection{References}{
\itemize{
\item Shavitt, I., & Segal, E. (2018). Regularization learning networks:
Deep learning for tabular datasets. \emph{Advances in Neural Information
Processing Systems}, 31, 1379-1389.
}
}
}
\keyword{internal}
