% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_models.R
\name{doc-tools}
\alias{doc-tools}
\alias{find_engine_files}
\alias{make_engine_list}
\alias{make_seealso_list}
\title{Tools for dynamically documenting packages}
\usage{
find_engine_files(mod)

make_engine_list(mod)

make_seealso_list(mod)
}
\arguments{
\item{mod}{A character string for the model file (e.g. "linear_reg")}
}
\value{
\code{make_engine_list()} returns a character string that creates a
bulletted list of links to more specific help files.

\code{make_seealso_list()} returns a formatted character string of links.

\code{find_engine_files()} returns a tibble.
}
\description{
These are internal functions used to create dynamic documentation in Rd files
based on which parsnip-related packages are loaded at any given time.

These functions can be used to make dynamic lists of documentation help
files. \pkg{parsnip} uses these, along with files contained in \code{man/rmd}
that have expanded documentation, for specific model/engine combinations.
\code{\link[=find_engine_files]{find_engine_files()}} looks for files that have the pattern
\verb{details_\{model\}_\{engine\}.Rd} to link to. These files are generated by files
named \verb{man/rmd/details_\{model\}_\{engine\}.Rmd}. \code{make_engine_list()} creates a
list seen at the top of the model Rd files while \code{make_seealso_list()}
populates the list seen in "See Also" below. See the details section.
}
\details{
The current \code{parsnip} documentation is \emph{dynamic}. There is a part at the
top of the Rd file that populates a list of engines that depends in what
packages are loaded at the time that the man file is loaded. For example, if
a side-package has a new engine for \code{linear_reg()}, the
\code{parsnip::linear_reg()} help can show a link to a detailed help page in the
side package.

To enable this, the process is to
\enumerate{
\item Create an engine-specific R file in the \code{R} directory with the name
\verb{\{model\}_\{engine\}.R} (e.g. \code{boost_tree_C5.0.R}). This has a small amount of
documentation as well as a directive
"\verb{@includeRmd man/rmd/\{model\}_\{engine\}.Rmd details}".
\item Copy the file in \code{parsnip} that is in \code{man/rmd/setup.Rmd} and put it in
the same place in your package.
\item Write your own \verb{man/rmd/\{model\}_\{engine\}.Rmd} file. This can include
packages that are not listed in the DESCRIPTION file. Those are only
required when the documentation file is created locally (probably using
\code{\link[devtools:document]{devtools::document()}}.
\item Run \code{\link[devtools:document]{devtools::document()}} so that the Rmd content is included in the
Rd file.
}
}
\examples{
find_engine_files("linear_reg")
cat(make_engine_list("linear_reg"))
}
\keyword{internal}
