% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_models.R
\name{doc-tools}
\alias{doc-tools}
\alias{find_engine_files}
\title{Tools for documenting packages}
\usage{
find_engine_files(mod)
}
\arguments{
\item{mod}{A character string for the model file}
}
\value{
\code{find_engine_files()} returns a character string that creates a
bulletted list of links to more specific help files.
}
\description{
Tools for documenting packages
}
\details{
This function can be used to make dynamic lists of documentation help
files. \pkg{parsnip} uses this along with files contained in \code{man/rmd} that
have expanded documentation for specific model/engine combinations.
\code{\link[=find_engine_files]{find_engine_files()}} looks for files that have the pattern
\verb{details_\{model\}_\{engine\}.Rd} to link to. These files are generated by files
named \verb{man/rmd/details_\{model\}_\{engine\}.Rmd}.
}
\examples{
cat(find_engine_files("linear_reg"))
}
\keyword{internal}
