% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helpers.R
\name{.extract_surv_time}
\alias{.extract_surv_time}
\title{Extract survival time}
\arguments{
\item{surv}{A single \code{\link[survival:Surv]{survival::Surv()}} object.}
}
\value{
A vector when the type is \code{"right"} or \code{"left"} and a tibble otherwise.
}
\description{
Extract the time component(s) from a \code{\link[survival:Surv]{survival::Surv()}} object.
}
