% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival-helpers.R
\name{.time_as_binary_event}
\alias{.time_as_binary_event}
\title{Convert survival objects to binary factors}
\arguments{
\item{surv}{A single \code{\link[survival:Surv]{survival::Surv()}} object.}

\item{eval_time}{A single numeric value for a landmark time.}
}
\value{
A two level factor.
}
\description{
For a given evaluation time, convert a \code{\link[survival:Surv]{survival::Surv()}} object to a binary
factor with levels \code{"event"} and \code{"non-event"}.
}
\details{
The following three cases can occur:
\itemize{
\item \strong{Definitive non-events}: event times (i.e., not censored) are less than
the evaluation time ("it hasn't happened yet")
\item \strong{Definitive events}: Observed times (censored or not) are greater than
the evaluation time ("it happens sometime after now").
\item \strong{Ambiguous outcomes}: Observed censored time is less than the evaluation
time ("maybe it happens, maybe not"). A missing value is returned for these
observations.
}
}
