% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit a Model Specification to a Dataset}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{An object of class \code{model_spec}}
}
\description{
\code{fit} will take a model specification, finalize the required code by substituting arguments, and execute the model fit routine.
}
