% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.model_spec}
\alias{fit.model_spec}
\alias{fit_xy}
\alias{fit_xy.model_spec}
\title{Fit a Model Specification to a Dataset}
\usage{
\method{fit}{model_spec}(object, formula = NULL, data = NULL,
  engine = object$engine, control = fit_control(), ...)

fit_xy(object, ...)

\method{fit_xy}{model_spec}(object, x = NULL, y = NULL,
  engine = object$engine, control = fit_control(), ...)
}
\arguments{
\item{object}{An object of class \code{model_spec}}

\item{formula}{An object of class "formula" (or one that can
be coerced to that class): a symbolic description of the model
to be fitted.}

\item{data}{Optional, depending on the interface (see Details
below). A data frame containing all relevant variables (e.g.
outcome(s), predictors, case weights, etc). Note: when needed, a
\emph{named argument} should be used.}

\item{engine}{A character string for the software that should
be used to fit the model. This is highly dependent on the type
of model (e.g. linear regression, random forest, etc.).}

\item{control}{A named list with elements \code{verbosity} and
\code{catch}. See \code{\link[=fit_control]{fit_control()}}.}

\item{...}{Not currently used; values passed here will be
ignored. Other options required to fit the model should be
passed using the \code{others} argument in the original model
specification.}

\item{x}{A matrix or data frame of predictors.}

\item{y}{A vector, matrix or data frame of outcome data.}
}
\value{
A \code{model_fit} object that contains several elements:
\itemize{
\item \code{lvl}: If the outcome is a factor, this contains
the factor levels at the time of model fitting.
\item \code{spec}: The model specification object
(\code{object} in the call to \code{fit})
\item \code{fit}: when the model is executed without error,
this is the model object. Otherwise, it is a \code{try-error}
object with the error message.
\item \code{preproc}: any objects needed to convert between
a formula and non-formula interface (such as the \code{terms}
object)
}
The return value will also have a class related to the fitted model (e.g.
\code{"_glm"}) before the base class of \code{"model_fit"}.
}
\description{
\code{fit} and \code{fit_xy} take a model specification, translate the required
code by substituting arguments, and execute the model fit
routine.
}
\details{
\code{fit} and \code{fit_xy} substitute the current arguments in the model
specification into the computational engine's code, checks them
for validity, then fits the model using the data and the
engine-specific code. Different model functions have different
interfaces (e.g. formula or \code{x}/\code{y}) and these functions translate
between the interface used when \code{fit} or \code{fit_xy} were invoked and the one
required by the underlying model.

When possible, these functions attempt to avoid making copies of the
data. For example, if the underlying model uses a formula and
\code{fit} is invoked, the original data are references
when the model is fit. However, if the underlying model uses
something else, such as \code{x}/\code{y}, the formula is evaluated and
the data are converted to the required format. In this case, any
calls in the resulting model objects reference the temporary
objects used to fit the model.
}
\examples{
# Although `glm` only has a formula interface, different
# methods for specifying the model can be used

library(dplyr)
data("lending_club")

lm_mod <- logistic_reg()

lm_mod <- logistic_reg()

using_formula <-
  lm_mod \%>\%
  fit(Class ~ funded_amnt + int_rate,
      data = lending_club,
      engine = "glm")

using_xy <-
  lm_mod \%>\%
  fit_xy(x = lending_club[, c("funded_amnt", "int_rate")],
         y = lending_club$Class,
         engine = "glm")

using_formula
using_xy
}
