% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{format-internals}
\alias{format-internals}
\alias{format_num}
\alias{format_class}
\alias{format_classprobs}
\alias{format_time}
\alias{format_survival}
\alias{format_linear_pred}
\alias{format_hazard}
\title{Internal functions that format predictions}
\usage{
format_num(x)

format_class(x)

format_classprobs(x)

format_time(x)

format_survival(x)

format_linear_pred(x)

format_hazard(x)
}
\arguments{
\item{x}{A data frame or vector (depending on the context and function).}
}
\value{
A tibble
}
\description{
These are used to ensure that we have appropriate column names inside of
tibbles.
}
\keyword{internal}
