% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_docs.R
\name{knit_engine_docs}
\alias{knit_engine_docs}
\title{Knit engine-specific documentation}
\usage{
knit_engine_docs(pattern = NULL)
}
\arguments{
\item{pattern}{A regular expression to specify which files to knit. The
default knits all engine documentation files.}
}
\value{
A tibble with column \code{file} for the file name and \code{result} (a
character vector that echos the output file name or, when there is
a failure, the error message).
}
\description{
Knit engine-specific documentation
}
\details{
This function will check whether the known parsnip extension packages,
engine specific packages, and a few other ancillary packages are installed.
Users will be prompted to install anything required to create the engine
documentation.
}
\keyword{internal}
