% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_reg.R
\name{logistic_reg}
\alias{logistic_reg}
\alias{logistic_reg.default}
\title{General Interface for Logistic Regression Models}
\usage{
logistic_reg(mode, ...)

\method{logistic_reg}{default}(mode = "classification", link = NULL,
  regularization = NULL, mixture = NULL, engine_args = list(), ...)
}
\arguments{
\item{mode}{A single character string for the type of model.
The only possible value for this model is "classification".}

\item{...}{Used for S3 method consistency. Any arguments passed to
the ellipses will result in an error. Use \code{engine_args} instead.}

\item{link}{A character string for the link function. Possible
values are "logit", "probit", "cauchit", "log" and "cloglog".}

\item{regularization}{An non-negative number representing the
total amount of regularization.}

\item{mixture}{A number between zero and one (inclusive) that
represents the proportion of regularization that is used for the
L2 penalty (i.e. weight decay, or ridge regression) versus L1
(the lasso).}

\item{engine_args}{A named list of arguments to be used by the
underlying models (e.g., \code{stats::glm},
\code{rstanarm::stan_glm}, etc.). These are not evaluated
until the model is fit and will be substituted into the model
fit expression.}
}
\description{
\code{logistic_reg} is a way to generate a \emph{specification} of a model
before fitting and allows the model to be created using
different packages in R, Stan, or via Spark. The main arguments for the
model are:
\itemize{
\item \code{link}: The link function.
\item \code{regularization}: The total amount of regularization
in the model. Note that this must be zero for some engines.
\item \code{mixture}: The proportion of L2 regularization in
the model. Note that this will be ignored for some engines.
}
These arguments are converted to their specific names at the
time that the model is fit. Other options and argument can be
set using the \code{engine_args} argument. If left to their defaults
here (\code{NULL}), the values are taken from the underlying model
functions.
}
\details{
The data given to the function are not saved and are only used
to determine the \emph{mode} of the model. For \code{logistic_reg},the
mode will always be "classification".

The model can be created using the \code{\link[=fit]{fit()}} function using the
following \emph{engines}:
\itemize{
\item \pkg{R}:  \code{"glm"} or \code{"glmnet"}
\item \pkg{Stan}:  \code{"rstanarm"}
\item \pkg{Spark}: \code{"spark"}
}
}
\examples{
logistic_reg()

# Parameters can be represented by a placeholder:
logistic_reg(link = "probit", regularization = varying())
}
\seealso{
\code{\link[=varying]{varying()}}, \code{\link[=fit]{fit()}}
}
