% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_db}
\alias{model_db}
\title{parsnip model specification database}
\value{
\item{model_db}{a data frame}
}
\description{
This is used in the RStudio add-in and captures information about mode
specifications in various R packages.
}
\examples{
\dontshow{if (!parsnip:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(model_db)
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
\keyword{internal}
