% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{multi_predict}
\alias{multi_predict}
\alias{multi_predict.default}
\title{Model predictions across many sub-models}
\usage{
multi_predict(object, ...)

\method{multi_predict}{default}(object, ...)
}
\arguments{
\item{object}{A \code{model_fit} object.}

\item{...}{Optional arguments to pass to \code{predict.model_fit(type = "raw")}
such as \code{type}.}
}
\value{
A tibble with the same number of rows as the data being predicted.
Mostly likely, there is a list-column named \code{.pred} that is a tibble with
multiple rows per sub-model.
}
\description{
For some models, predictions can be made on sub-models in the model object.
}
\keyword{internal}
