% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbor.R
\name{nearest_neighbor}
\alias{nearest_neighbor}
\title{K-nearest neighbor}
\usage{
nearest_neighbor(
  mode = "unknown",
  neighbors = NULL,
  weight_func = NULL,
  dist_power = NULL
)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible values for this model are \code{"unknown"}, \code{"regression"}, or
\code{"classification"}.}

\item{neighbors}{A single integer for the number of neighbors
to consider (often called \code{k}). For \pkg{kknn}, a value of 5
is used if \code{neighbors} is not specified.}

\item{weight_func}{A \emph{single} character for the type of kernel function used
to weight distances between samples. Valid choices are: \code{"rectangular"},
\code{"triangular"}, \code{"epanechnikov"}, \code{"biweight"}, \code{"triweight"},
\code{"cos"}, \code{"inv"}, \code{"gaussian"}, \code{"rank"}, or \code{"optimal"}.}

\item{dist_power}{A single number for the parameter used in
calculating Minkowski distance.}
}
\description{
\code{nearest_neighbor()} defines a model that uses the K most similar data points
from the training set to predict new samples.

There are different ways to fit this model. See the engine-specific pages
for more details:

\Sexpr[stage=render,results=rd]{parsnip:::find_engine_files("nearest_neighbor")}

More information on how \code{parsnip} is used for modeling is at
\url{https://www.tidymodels.org}.
}
\details{
This function only defines what \emph{type} of model is being fit. Once an engine
is specified, the \emph{method} to fit the model is also defined.

The model is not trained or fit until the \code{\link[=fit.model_spec]{fit.model_spec()}} function is used
with the data.
}
\examples{
show_engines("nearest_neighbor")

nearest_neighbor(neighbors = 11)

}
\references{
\url{https://www.tidymodels.org},
\href{https://tmwr.org}{\emph{Tidy Models with R}}
}
\seealso{
\code{\link[=fit.model_spec]{fit.model_spec()}}, \code{\link[=set_engine]{set_engine()}}, \code{\link[=update]{update()}},
\code{\link[=details_nearest_neighbor_kknn]{kknn engine details}}
}
