% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel.R
\name{null_model}
\alias{null_model}
\alias{null_model.default}
\alias{predict.null_model}
\alias{print.null_model}
\title{Fit a simple, non-informative model}
\usage{
null_model(x, ...)

\method{null_model}{default}(x = NULL, y, ...)

\method{print}{null_model}(x, ...)

\method{predict}{null_model}(object, newdata = NULL, type = NULL, ...)
}
\arguments{
\item{x}{An optional matrix or data frame of predictors. These values are
not used in the model fit}

\item{\dots}{Optional arguments (not yet used)}

\item{y}{A numeric vector (for regression) or factor (for classification) of
outcomes}

\item{object}{An object of class \code{null_model}}

\item{newdata}{A matrix or data frame of predictors (only used to determine
the number of predictions to return)}

\item{type}{Either "raw" (for regression), "class" or "prob" (for
classification)}
}
\value{
The output of \code{null_model} is a list of class \code{null_model}
with elements \item{call }{the function call} \item{value }{the mean of
\code{y} or the most prevalent class} \item{levels }{when \code{y} is a
factor, a vector of levels. \code{NULL} otherwise} \item{pct }{when \code{y}
is a factor, a data frame with a column for each class (\code{NULL}
otherwise). The column for the most prevalent class has the proportion of
the training samples with that class (the other columns are zero). } \item{n
}{the number of elements in \code{y}}

\code{predict.null_model} returns a either a factor or numeric vector
depending on the class of \code{y}. All predictions are always the same.
}
\description{
Fit a single mean or largest class model
}
\details{
\code{null_model} emulates other model building functions, but returns the
simplest model possible given a training set: a single mean for numeric
outcomes and the most prevalent class for factor outcomes. When class
probabilities are requested, the percentage of the training set samples with
the most prevalent class is returned.
}
\examples{

outcome <- factor(sample(letters[1:2],
                         size = 100,
                         prob = c(.1, .9),
                         replace = TRUE))
useless <- null_model(y = outcome)
useless
predict(useless, matrix(NA, nrow = 10))

}
\keyword{models}
