% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel.R
\name{null_model}
\alias{null_model}
\title{Null model}
\usage{
null_model(mode = "classification")
}
\arguments{
\item{mode}{A single character string for the prediction outcome mode.
Possible values for this model are "unknown", "regression", or
"classification".}
}
\description{
\code{null_model()} defines a simple, non-informative model. It doesn't have any
main arguments.
}
\details{
The model can be created using the \code{fit()} function using the
following \emph{engines}:
\itemize{
\item \pkg{R}:  \code{"parsnip"}
}
}
\section{Engine Details}{
Engines may have pre-set default arguments when executing the model fit
call. For this type of model, the template of the fit calls are below:
\subsection{parsnip}{\if{html}{\out{<div class="sourceCode r">}}\preformatted{null_model() \%>\% 
  set_engine("parsnip") \%>\% 
  set_mode("regression") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Model Specification (regression)
## 
## Computational engine: parsnip 
## 
## Model fit template:
## nullmodel(x = missing_arg(), y = missing_arg())
}\if{html}{\out{<div class="sourceCode r">}}\preformatted{null_model() \%>\% 
  set_engine("parsnip") \%>\% 
  set_mode("classification") \%>\% 
  translate()
}\if{html}{\out{</div>}}\preformatted{## Model Specification (classification)
## 
## Computational engine: parsnip 
## 
## Model fit template:
## nullmodel(x = missing_arg(), y = missing_arg())
}
}
}

\examples{
null_model(mode = "regression")
}
\seealso{
\code{\link[=fit.model_spec]{fit.model_spec()}}
}
