% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmodel.R
\name{null_model}
\alias{null_model}
\title{Null model}
\usage{
null_model(mode = "classification", engine = "parsnip")
}
\arguments{
\item{mode}{A single character string for the type of model. The only
possible values for this model are \code{"regression"} and \code{"classification"}.}

\item{engine}{A single character string specifying what computational engine
to use for fitting. Possible engines are listed below. The default for this
model is \code{"parsnip"}.}
}
\description{
\code{null_model()} defines a simple, non-informative model. It doesn't have any
main arguments. This function can fit classification and regression models.
}
\section{Engine Details}{
Engines may have pre-set default arguments when executing the model fit
call. For this type of model, the template of the fit calls are below:
\subsection{parsnip}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{null_model() |> 
  set_engine("parsnip") |> 
  set_mode("regression") |> 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Null Model Specification (regression)
## 
## Computational engine: parsnip 
## 
## Model fit template:
## parsnip::nullmodel(x = missing_arg(), y = missing_arg())
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{null_model() |> 
  set_engine("parsnip") |> 
  set_mode("classification") |> 
  translate()
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode">}}\preformatted{## Null Model Specification (classification)
## 
## Computational engine: parsnip 
## 
## Model fit template:
## parsnip::nullmodel(x = missing_arg(), y = missing_arg())
}\if{html}{\out{</div>}}
}
}

\examples{
\dontshow{if (!parsnip:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
null_model(mode = "regression")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=fit.model_spec]{fit.model_spec()}}
}
