% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_reg.R
\name{ordinal_reg}
\alias{ordinal_reg}
\title{Ordinal regression}
\usage{
ordinal_reg(
  mode = "classification",
  ordinal_link = NULL,
  odds_link = NULL,
  engine = "polr"
)
}
\arguments{
\item{mode}{A single character string for the prediction outcome mode. The
only possible value for this model is "classification".}

\item{ordinal_link}{The ordinal link function.}

\item{odds_link}{The odds or probability link function.}

\item{engine}{A single character string specifying what computational engine
to use for fitting. Possible engines are listed below. The default for this
model is \code{"polr"}.}
}
\description{
\code{ordinal_reg()} defines a generalized linear model that predicts an ordinal
outcome. This function can fit classification models.

\Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("ordinal_reg")}

More information on how \pkg{parsnip} is used for modeling is at
\url{https://www.tidymodels.org/}.
}
\details{
This function only defines what \emph{type} of model is being fit. Once an engine
is specified, the \emph{method} to fit the model is also defined. See
\code{\link[=set_engine]{set_engine()}} for more on setting the engine, including how to set engine
arguments.

The model is not trained or fit until the \code{\link[=fit.model_spec]{fit()}} function is used
with the data.

Each of the arguments in this function other than \code{mode} and \code{engine} are
captured as \link[rlang:topic-quosure]{quosures}. To pass values
programmatically, use the \link[rlang:injection-operator]{injection operator} like so:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{value <- 1
ordinal_reg(argument = !!value)
}\if{html}{\out{</div>}}

Ordinal regression models include cumulative, sequential, and
adjacent structures.
}
\examples{
\dontshow{if (!parsnip:::is_cran_check()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
show_engines("ordinal_reg")

ordinal_reg(mode = "classification")
\dontshow{\}) # examplesIf}
}
\references{
\url{https://www.tidymodels.org}, \href{https://www.tmwr.org/}{\emph{Tidy Modeling with R}}, \href{https://www.tidymodels.org/find/parsnip/}{searchable table of parsnip models}
}
\seealso{
\Sexpr[stage=render,results=rd]{parsnip:::make_seealso_list("ordinal_reg")}
}
\keyword{internal}
