% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_class.R, R/predict_classprob.R,
%   R/predict_interval.R, R/predict_num.R
\name{predict_class.model_fit}
\alias{predict_class.model_fit}
\alias{predict_class}
\alias{predict_classprob.model_fit}
\alias{predict_classprob}
\alias{predict_confint.model_fit}
\alias{predict_confint}
\alias{predict_predint.model_fit}
\alias{predict_predint}
\alias{predict_num.model_fit}
\alias{predict_num}
\title{Other predict methods.}
\usage{
\method{predict_class}{model_fit}(object, new_data, ...)

predict_class(object, ...)

\method{predict_classprob}{model_fit}(object, new_data, ...)

predict_classprob(object, ...)

\method{predict_confint}{model_fit}(object, new_data, level = 0.95,
  std_error = FALSE, ...)

predict_confint(object, ...)

\method{predict_predint}{model_fit}(object, new_data, level = 0.95,
  std_error = FALSE, ...)

predict_predint(object, ...)

\method{predict_num}{model_fit}(object, new_data, ...)

predict_num(object, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{...}{Ignored. To pass arguments to pass to the underlying
function when \code{predict.model_fit(type = "raw")},
use the \code{opts} argument.}

\item{level}{A single numeric value between zero and one for the
interval estimates.}

\item{std_error}{A single logical for wether the standard error should be
returned (assuming that the model can compute it).}
}
\description{
These are internal functions not meant to be directly called by the user.
}
\keyword{internal}
