% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.model_fit}
\alias{predict.model_fit}
\title{Regression Model Predictions}
\usage{
\method{predict}{model_fit}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{newdata}{A rectangular data object, such as a data frame.}

\item{...}{Arguments to pass to other methods (not currently used).}
}
\value{
An unnamed numeric vector of predictions.
}
\description{
Apply a model to create predictions when the outcome is a simple
number.
}
