% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/predict_raw.R
\name{predict.model_fit}
\alias{predict.model_fit}
\alias{predict_raw.model_fit}
\alias{predict_raw}
\title{Model predictions}
\usage{
\method{predict}{model_fit}(object, new_data, type = NULL,
  opts = list(), ...)

\method{predict_raw}{model_fit}(object, new_data, opts = list(), ...)

predict_raw(object, ...)
}
\arguments{
\item{object}{An object of class \code{model_fit}}

\item{new_data}{A rectangular data object, such as a data frame.}

\item{type}{A single character value or \code{NULL}. Possible values
are "numeric", "class", "probs", "conf_int", "pred_int", or
"raw". When \code{NULL}, \code{predict} will choose an appropriate value
based on the model's mode.}

\item{opts}{A list of optional arguments to the underlying
predict function that will be used when \code{type = "raw"}. The
list should not include options for the model object or the
new data being predicted.}

\item{...}{Ignored. To pass arguments to pass to the underlying
function when \code{predict.model_fit(type = "raw")},
use the \code{opts} argument.}
}
\value{
With the exception of \code{type = "raw"}, the results of
\code{predict.model_fit} will be a tibble as many rows in the output
as there are rows in \code{new_data} and the column names will be
predictable.

For numeric results with a single outcome, the tibble will have
a \code{.pred} column and \code{.pred_Yname} for multivariate results.

For hard class predictions, the column is named \code{.pred_class}
and, when \code{type = "prob"}, the columns are \code{.pred_classlevel}.

\code{type = "conf_int"} and \code{type = "pred_int"} return tibbles with
columns \code{.pred_lower} and \code{.pred_upper} with an attribute for
the confidence level. In the case where intervals can be
produces for class probabilities (or other non-scalar outputs),
the columns will be named \code{.pred_lower_classlevel} and so on.

Using \code{type = "raw"} with \code{predict.model_fit} (or using
\code{predict_raw}) will return the unadulterated results of the
prediction function.

In the case of Spark-based models, since table columns cannot
contain dots, the same convention is used except 1) no dots
appear in names and 2) vectors are never returned but
type-specific prediction functions.
}
\description{
Apply a model to create different types of predictions.
\code{predict} can be used for all types of models and used the
"type" argument for more specificity.
}
\details{
If "type" is not supplied to \code{predict}, then a choice
is made (\code{type = "numeric"} for regression models and
\code{type = "class"} for classification).

\code{predict} is designed to provide a tidy result (see "Value"
section below) in a tibble output format.

When using \code{type = "conf_int"} and \code{type = "pred_int"}, the options
\code{level} and \code{std_error} can be used. The latter is a logical for an
extra column of standard error values (if available).
}
\examples{
library(dplyr)

lm_model <-
  linear_reg() \%>\%
  fit(mpg ~ ., data = mtcars \%>\% slice(11:32), engine = "lm")

pred_cars <-
  mtcars \%>\%
  slice(1:10) \%>\%
  select(-mpg)

predict(lm_model, pred_cars)

predict(
  lm_model,
  pred_cars,
  type = "conf_int",
  level = 0.90
)

predict(
  lm_model,
  pred_cars,
  type = "raw",
  opts = list(type = "terms")
)
}
