% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R, R/varying.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{fit}
\alias{fit_xy}
\alias{tidy}
\alias{glance}
\alias{augment}
\alias{varying_args}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{fit}}, \code{\link[generics]{fit_xy}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}, \code{\link[generics]{varying_args}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

