% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirects.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{fit}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

