% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_pkgs.R
\name{req_pkgs}
\alias{req_pkgs}
\title{Determine required packages for a model}
\usage{
req_pkgs(x, ...)
}
\arguments{
\item{x}{A \link[=model_spec]{model specification} or \link[=model_fit]{fit}.}

\item{...}{Not used.}
}
\value{
A character string of package names (if any).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function has been deprecated in favor of \code{required_pkgs()}.
}
