% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engines.R
\name{show_engines}
\alias{show_engines}
\title{Display available engines for a model}
\usage{
show_engines(x)
}
\arguments{
\item{x}{The name of a \code{parsnip} model (e.g., "linear_reg", "mars", etc.)}
}
\value{
A tibble.
show_engines("linear_reg")
}
\description{
The possible engines for a model may depend on what packages are loaded. There
are some \code{parsnip}-adjacent packages that only add engines to existing models.
}
