% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_reg.R
\name{survival_reg}
\alias{survival_reg}
\title{Parametric survival regression}
\usage{
survival_reg(mode = "censored regression", engine = "survival", dist = NULL)
}
\arguments{
\item{mode}{A single character string for the prediction outcome mode.
The only possible value for this model is "censored regression".}

\item{engine}{A single character string specifying what computational engine
to use for fitting.}

\item{dist}{A character string for the probability distribution of the
outcome. The default is "weibull".}
}
\description{
\code{survival_reg()} defines a parametric survival model. This function can fit
censored regression models.

\Sexpr[stage=render,results=rd]{parsnip:::make_engine_list("survival_reg")}

More information on how \pkg{parsnip} is used for modeling is at
\url{https://www.tidymodels.org/}.
}
\details{
This function only defines what \emph{type} of model is being fit. Once an engine
is specified, the \emph{method} to fit the model is also defined.

The model is not trained or fit until the \code{\link[=fit.model_spec]{fit()}} function is used
with the data.

Since survival models typically involve censoring (and require the use of
\code{\link[survival:Surv]{survival::Surv()}} objects), the \code{\link[=fit.model_spec]{fit.model_spec()}} function will require that the
survival model be specified via the formula interface.
}
\examples{
show_engines("survival_reg")

survival_reg(mode = "censored regression", dist = "weibull")
}
\references{
\url{https://www.tidymodels.org}, \href{https://www.tmwr.org/}{\emph{Tidy Modeling with R}}, \href{https://www.tidymodels.org/find/parsnip/}{searchable table of parsnip models}
}
\seealso{
\Sexpr[stage=render,results=rd]{parsnip:::make_seealso_list("survival_reg")}
}
\keyword{internal}
