% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_linear.R
\name{svm_linear}
\alias{svm_linear}
\title{Linear support vector machines}
\usage{
svm_linear(mode = "unknown", cost = NULL, margin = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible values for this model are "unknown", "regression", or
"classification".}

\item{cost}{A positive number for the cost of predicting a sample within
or on the wrong side of the margin}

\item{margin}{A positive number for the epsilon in the SVM insensitive
loss function (regression only)}
}
\description{
\code{svm_linear()} defines a support vector machine model. For classification,
these models try to maximize the width of the margin between classes.
For regression, the model optimizes a robust loss function that is only
affected by very large model residuals.

This SVM model uses a linear function to create the decision boundary or
regression line.

There are different ways to fit this model. Information about the available
\emph{engines} that can be used for fitting at:

\Sexpr[stage=render,results=rd]{parsnip:::find_engine_files("svm_linear")}

More information on how \code{parsnip} is used for model is at
\url{https://www.tidymodels.org}.
}
\details{
This function only defines what \emph{type} of model is being fit. Once an engine
is specified, the \emph{method} to fit the model is also defined.

The model is not trained or fit until the \code{\link[=fit.model_spec]{fit.model_spec()}} function is used
with the data.
}
\examples{
show_engines("svm_linear")

svm_linear(mode = "classification")
}
\references{
\url{https://www.tidymodels.org},
\href{https://tmwr.org}{\emph{Tidy Models with R}}
}
\seealso{
\code{\link[=fit.model_spec]{fit.model_spec()}}, \code{\link[=set_engine]{set_engine()}}, \code{\link[=update]{update()}},
\code{\link[=details_svm_linear_LiblineaR]{LiblineaR engine details}},
\code{\link[=details_svm_linear_kernlab]{kernlab engine details}}
}
