% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Resolve a Model Specification for a Computational Engine}
\usage{
translate(x, ...)
}
\arguments{
\item{x}{A model specification.}

\item{...}{Not currently used.}
}
\description{
\code{translate} will translate a model specification into a code
object that is specific to a particular engine (e.g. R package).
It translates generic parameters to their counterparts.
}
\details{
\code{translate} produces a \emph{template} call that lacks the specific
argument values (such as \code{data}, etc). These are filled in once
\code{fit} is called with the specifics of the data for the model.
The call may also include \code{varying} arguments if these are in
the specification.

It does contain the resolved argument names that are specific to
the model fitting function/engine.

This function can be useful when you need to understand how
\code{parsnip} goes from a generic model specific to a model fitting
function.
}
\examples{
lm_spec <- linear_reg(penalty = 0.01)

# `penalty` is tranlsated to `lambda`
translate(lm_spec, engine = "glmnet")

# `penalty` not applicable for this model.
translate(lm_spec, engine = "lm")

# `penalty` is tranlsated to `reg_param`
translate(lm_spec, engine = "spark")

# with a placeholder for an unknown argument value:
translate(linear_reg(mixture = varying()), engine = "glmnet")

}
