% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine_docs.R
\name{update_model_info_file}
\alias{update_model_info_file}
\title{Save information about models}
\usage{
update_model_info_file(path = "inst/models.tsv")
}
\arguments{
\item{path}{A character string for the location of the tab delimited file.}
}
\description{
This function writes a tab delimited file to the package to capture
information about the known models. This information includes packages in
the tidymodels GitHub repository as well as packages that are known to work
well with tidymodels packages (e.g. not only \pkg{parsnip} but also
\pkg{tune}, etc.). There may be more model definitions in other extension
packages that are not included here.

These data are used to document engines for each model function man page.
}
\details{
See our
\href{https://tidymodels.github.io/model-implementation-principles/}{model implementation guidelines}
on best practices for modeling and modeling packages.

It is highly recommended that the known parsnip extension packages are loaded.
The unexported \pkg{parsnip} function \code{extensions()} will list these.
}
\keyword{internal}
