% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varying.R
\name{varying}
\alias{varying}
\title{A placeholder function for argument values}
\usage{
varying()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link[=varying]{varying()}} is used when a parameter will be specified at a later date.
}
\keyword{internal}
