% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_quantiles.R
\name{vec_quantiles}
\alias{vec_quantiles}
\title{A vector containing sets of quantiles}
\usage{
vec_quantiles(values, quantile_levels = NULL)
}
\arguments{
\item{values}{A data.frame/matrix/vector of values. If a named data.frame,
the column names will be used as the \code{quantile_levels} if those are missing.}

\item{quantile_levels}{A vector of probabilities corresponding to \code{values}.
May be \code{NULL} if \code{values} is a named data.frame.}
}
\description{
A vector containing sets of quantiles
}
\examples{
preds <- vec_quantiles(list(1:4, 8:11), c(.2, .4, .6, .8))

vec_quantiles(1:4, 1:4 / 5)
}
