% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_quantiles.R
\name{vec_quantiles}
\alias{vec_quantiles}
\title{A vector containing sets of quantiles}
\usage{
vec_quantiles(values, quantile_levels = double())
}
\arguments{
\item{values}{A matrix of values. Each column should correspond to one of
the quantile levels.}

\item{quantile_levels}{A vector of probabilities corresponding to \code{values}.}
}
\description{
A vector containing sets of quantiles
}
\examples{
vec_quantiles(matrix(rnorm(20), 5), c(.2, .4, .6, .8))
}
